Split up seed data
