+++ bin/counts.ts
@@ -1,9 +1,7 @@
-import {
-  allAlbums,
-  allPerformances,
-  allShows,
-  allSongs,
-} from "../src/drizzle/seeds";
+import { allAlbums } from "@/drizzle/data/albums";
+import { allPerformances } from "@/drizzle/data/performances";
+import { allShows } from "@/drizzle/data/shows";
+import { allSongs } from "@/drizzle/data/songs";
 
 console.log(`Albums:       ${allAlbums.length}`);
 console.log(`Songs:        ${allSongs.length}`);
