+++ src/drizzle/data/albums.ts
@@ -1,0 +1,212 @@
+import { Album } from "../schema";
+
+// =============================================================================
+// ================================ ALBUMS =====================================
+// =============================================================================
+
+// Ordered by release date
+export const seedAlbums = {
+  tbb: {
+    title: "12 Bar Bruise",
+    slug: "12-bar-bruise",
+    releaseDate: "2012-09-07",
+    id: "5bc6a134-a0dc-4184-9b64-3f7bdd003a92",
+    imageUrl: "https://f4.bcbits.com/img/a3529909906_16.jpg",
+    bandcampAlbumId: "2765388374",
+  },
+  elts: {
+    title: "Eyes Like the Sky",
+    slug: "eyes-like-the-sky",
+    releaseDate: "2013-02-22",
+    id: "bee4a607-7f13-438a-af17-de6029e1ffbc",
+    imageUrl: "https://f4.bcbits.com/img/a1198126189_16.jpg",
+    bandcampAlbumId: "781831684",
+  },
+  fafyl: {
+    title: "Float Along - Fill Your Lungs",
+    slug: "float-along-fill-your-lungs",
+    releaseDate: "2013-09-27",
+    id: "55154b57-7925-408c-af21-c94fe06b3d62",
+    imageUrl: "https://f4.bcbits.com/img/a3907213656_16.jpg",
+    bandcampAlbumId: "1510489151",
+  },
+  oddments: {
+    title: "Oddments",
+    slug: "oddments",
+    releaseDate: "2014-03-07",
+    id: "7f1d12b4-d57a-42e8-9aae-57082cab5633",
+    imageUrl: "https://f4.bcbits.com/img/a0348996479_16.jpg",
+    bandcampAlbumId: "3657754108",
+  },
+  mindFuzz: {
+    title: "I'm In Your Mind Fuzz",
+    slug: "im-in-your-mind-fuzz",
+    releaseDate: "2014-10-31",
+    id: "29ae645d-1487-4963-9605-25a24904c205",
+    imageUrl: "https://f4.bcbits.com/img/a3937834400_16.jpg",
+    bandcampAlbumId: "3101239241",
+  },
+  quarters: {
+    title: "Quarters",
+    slug: "quarters",
+    releaseDate: "2015-05-12",
+    id: "5ba9fb14-390a-4f14-a141-21b6fb66adac",
+    imageUrl: "https://f4.bcbits.com/img/a3796441552_16.jpg",
+    bandcampAlbumId: "1306577560",
+  },
+  pmdb: {
+    title: "Paper Mâché Dream Balloon",
+    slug: "paper-mache-dream-balloon",
+    releaseDate: "2015-11-13",
+    id: "1d5545eb-e953-41a7-be6d-a3dbb9f5d8e7",
+    imageUrl: "https://f4.bcbits.com/img/a2280704800_16.jpg",
+    bandcampAlbumId: "1762740948",
+  },
+  nonagon: {
+    title: "Nonagon Infinity",
+    slug: "nonagon-infinity",
+    releaseDate: "2016-04-29",
+    id: "8a1332c8-c7d2-481c-9551-afa8bab79bd6",
+    imageUrl: "https://f4.bcbits.com/img/a1566708845_16.jpg",
+    bandcampAlbumId: "87057479",
+  },
+  fmb: {
+    title: "Flying Microtonal Banana",
+    slug: "flying-microtonal-banana",
+    releaseDate: "2017-02-24",
+    id: "003314d7-fe25-4742-8b47-cb52f7966bc2",
+    imageUrl: "https://f4.bcbits.com/img/a2731568276_16.jpg",
+    bandcampAlbumId: "114372498",
+  },
+  motu: {
+    title: "Murder of the Universe",
+    slug: "murder-of-the-universe",
+    releaseDate: "2017-06-23",
+    id: "45419e68-2701-4221-8c6d-ee29f66523de",
+    imageUrl: "https://f4.bcbits.com/img/a3406870115_16.jpg",
+    bandcampAlbumId: "1787442871",
+  },
+  sketches: {
+    title: "Sketches of Brunswick East",
+    slug: "sketches-of-brunswick-east",
+    releaseDate: "2017-08-18",
+    id: "a8fcce86-fab9-44d1-b3c2-1c6b1297ab2e",
+    imageUrl: "https://f4.bcbits.com/img/a4206546710_16.jpg",
+    bandcampAlbumId: "1255749163",
+  },
+  poly: {
+    title: "Polygondwanaland",
+    slug: "polygondwanaland",
+    releaseDate: "2017-11-17",
+    id: "03c0b2d8-626c-4c1c-b0f9-5c730477caab",
+    imageUrl: "https://f4.bcbits.com/img/a1550557789_16.jpg",
+    bandcampAlbumId: "1017207616",
+  },
+  gumbootSoup: {
+    title: "Gumboot Soup",
+    slug: "gumboot-soup",
+    releaseDate: "2017-12-31",
+    id: "08816b5c-5e1a-4aca-b07c-0a1b1008fabe",
+    imageUrl: "https://f4.bcbits.com/img/a1481325267_16.jpg",
+    bandcampAlbumId: "2728208210",
+  },
+  fff: {
+    title: "Fishing For Fishies",
+    slug: "fishing-for-fishies",
+    releaseDate: "2019-04-26",
+    id: "bc67b3f7-af1d-443f-9d28-1c99f6d4aa60",
+    imageUrl: "https://f4.bcbits.com/img/a0470322543_16.jpg",
+    bandcampAlbumId: "1507780262",
+  },
+  itrn: {
+    title: "Infest the Rats' Nest",
+    slug: "infest-the-rats-nest",
+    releaseDate: "2019-08-16",
+    id: "e889bd13-f73f-40fc-8ea9-17a6c3e9ad26",
+    imageUrl: "https://f4.bcbits.com/img/a0402916090_16.jpg",
+    bandcampAlbumId: "3925296262",
+  },
+  kg: {
+    title: "K.G.",
+    slug: "kg",
+    releaseDate: "2020-11-20",
+    id: "0c1f66d2-4ebc-499d-9bd7-c1ea34e4fbc0",
+    imageUrl: "https://f4.bcbits.com/img/a3546319741_16.jpg",
+    bandcampAlbumId: "927421026",
+  },
+  lw: {
+    title: "L.W.",
+    slug: "lw",
+    releaseDate: "2021-02-26",
+    id: "0b3e78ca-cfd3-4e74-8c47-5495271d0cc2",
+    imageUrl: "https://f4.bcbits.com/img/a2018491224_16.jpg",
+    bandcampAlbumId: "94095409",
+  },
+  bf3k: {
+    title: "Butterfly 3000",
+    slug: "butterfly-3000",
+    releaseDate: "2021-06-11",
+    id: "40fdce93-4438-4700-8721-9fade20341a0",
+    imageUrl: "https://f4.bcbits.com/img/a3009849752_16.jpg",
+    bandcampAlbumId: "3244133050",
+  },
+  og: {
+    title: "Omnium Gatherum",
+    slug: "omnium-gatherum",
+    releaseDate: "2022-04-22",
+    id: "560b0194-19ae-4bc9-aa24-0c2340596af7",
+    imageUrl: "https://f4.bcbits.com/img/a2638998972_16.jpg",
+    bandcampAlbumId: "129981996",
+  },
+  iceDeath: {
+    title: "Ice, Death, Planets, Lungs, Mushrooms and Lava",
+    slug: "ice-death-planets-lungs-mushrooms-and-lava",
+    releaseDate: "2022-07-10",
+    id: "46f925f1-dd7f-4966-8c02-3562e32f22b5",
+    imageUrl: "https://f4.bcbits.com/img/a3317842641_16.jpg",
+    bandcampAlbumId: "980855912",
+  },
+  laminatedDenim: {
+    title: "Laminated Denim",
+    slug: "laminate-denim",
+    releaseDate: "2022-10-12",
+    id: "7ceb3610-4aac-41cb-9a2f-4b6633dfad9b",
+    imageUrl: "https://f4.bcbits.com/img/a3665205369_16.jpg",
+    bandcampAlbumId: "177111450",
+  },
+  changes: {
+    title: "Changes",
+    slug: "changes",
+    releaseDate: "2022-10-28",
+    id: "957eae4b-eec8-45de-abcc-8902cda88fd3",
+    imageUrl: "https://f4.bcbits.com/img/a1890146474_16.jpg",
+    bandcampAlbumId: "2202545504",
+  },
+  petro: {
+    title:
+      "PetroDragonic Apocalypse; or, Dawn of Eternal Night: An Annihilation of Planet Earth and the Beginning of Merciless Damnation",
+    slug: "petro",
+    releaseDate: "2023-06-16",
+    id: "f43bb6f3-6b08-4687-a39e-a4f379325f93",
+    imageUrl: "https://f4.bcbits.com/img/a2805471381_16.jpg",
+    bandcampAlbumId: "3699259806",
+  },
+  tsc: {
+    title: "The Silver Cord",
+    slug: "the-silver-cord",
+    releaseDate: "2023-10-27",
+    id: "e212c176-b78f-4b0b-93a0-651ca99591e2",
+    imageUrl: "https://f4.bcbits.com/img/a0297097933_16.jpg",
+    bandcampAlbumId: "2151949833",
+  },
+  b741: {
+    title: "Flight b741",
+    slug: "flight-b741",
+    releaseDate: "2024-08-09",
+    id: "3e106926-3ea8-4742-87b4-7ea809799370",
+    imageUrl: "https://f4.bcbits.com/img/a3548545002_16.jpg",
+    bandcampAlbumId: "104852051",
+  },
+} as const satisfies Record<string, Omit<Album, "createdAt">>;
+
+export const allAlbums = Object.values(seedAlbums) as Album[];
