+++ src/drizzle/data/performances.ts
@@ -19,10 +19,6 @@
 //  option is to go to https://duck.com/?q=uuid and copy the generated UUID from
 //  there.)
 //
-//  I've taken the image URLs from Bandcamp album pages. For those shows that do
-//  not have official albums on Bandcamp, I either cropped squares from the show
-//  poster or leave the show with no image.
-//
 //  You can find the Spotify/Bandcamp/YouTube IDs by going to the relevant video
 //  or song on that service, generating an iframe snippet, and finding the ID in
 //  the generated HTML.
@@ -32,2485 +28,12 @@
 //
 // =============================================================================
 
-import { Album, Performance, Show, Song } from "./schema";
+import { Performance } from "../schema";
+import { seedShows } from "./shows";
+import { seedSongs } from "./songs";
 
 // =============================================================================
-// ================================ ALBUMS =====================================
-// =============================================================================
-
-// Ordered by release date
-export const seedAlbums = {
-  tbb: {
-    title: "12 Bar Bruise",
-    slug: "12-bar-bruise",
-    releaseDate: "2012-09-07",
-    id: "5bc6a134-a0dc-4184-9b64-3f7bdd003a92",
-    imageUrl: "https://f4.bcbits.com/img/a3529909906_16.jpg",
-    bandcampAlbumId: "2765388374",
-  },
-  elts: {
-    title: "Eyes Like the Sky",
-    slug: "eyes-like-the-sky",
-    releaseDate: "2013-02-22",
-    id: "bee4a607-7f13-438a-af17-de6029e1ffbc",
-    imageUrl: "https://f4.bcbits.com/img/a1198126189_16.jpg",
-    bandcampAlbumId: "781831684",
-  },
-  fafyl: {
-    title: "Float Along - Fill Your Lungs",
-    slug: "float-along-fill-your-lungs",
-    releaseDate: "2013-09-27",
-    id: "55154b57-7925-408c-af21-c94fe06b3d62",
-    imageUrl: "https://f4.bcbits.com/img/a3907213656_16.jpg",
-    bandcampAlbumId: "1510489151",
-  },
-  oddments: {
-    title: "Oddments",
-    slug: "oddments",
-    releaseDate: "2014-03-07",
-    id: "7f1d12b4-d57a-42e8-9aae-57082cab5633",
-    imageUrl: "https://f4.bcbits.com/img/a0348996479_16.jpg",
-    bandcampAlbumId: "3657754108",
-  },
-  mindFuzz: {
-    title: "I'm In Your Mind Fuzz",
-    slug: "im-in-your-mind-fuzz",
-    releaseDate: "2014-10-31",
-    id: "29ae645d-1487-4963-9605-25a24904c205",
-    imageUrl: "https://f4.bcbits.com/img/a3937834400_16.jpg",
-    bandcampAlbumId: "3101239241",
-  },
-  quarters: {
-    title: "Quarters",
-    slug: "quarters",
-    releaseDate: "2015-05-12",
-    id: "5ba9fb14-390a-4f14-a141-21b6fb66adac",
-    imageUrl: "https://f4.bcbits.com/img/a3796441552_16.jpg",
-    bandcampAlbumId: "1306577560",
-  },
-  pmdb: {
-    title: "Paper Mâché Dream Balloon",
-    slug: "paper-mache-dream-balloon",
-    releaseDate: "2015-11-13",
-    id: "1d5545eb-e953-41a7-be6d-a3dbb9f5d8e7",
-    imageUrl: "https://f4.bcbits.com/img/a2280704800_16.jpg",
-    bandcampAlbumId: "1762740948",
-  },
-  nonagon: {
-    title: "Nonagon Infinity",
-    slug: "nonagon-infinity",
-    releaseDate: "2016-04-29",
-    id: "8a1332c8-c7d2-481c-9551-afa8bab79bd6",
-    imageUrl: "https://f4.bcbits.com/img/a1566708845_16.jpg",
-    bandcampAlbumId: "87057479",
-  },
-  fmb: {
-    title: "Flying Microtonal Banana",
-    slug: "flying-microtonal-banana",
-    releaseDate: "2017-02-24",
-    id: "003314d7-fe25-4742-8b47-cb52f7966bc2",
-    imageUrl: "https://f4.bcbits.com/img/a2731568276_16.jpg",
-    bandcampAlbumId: "114372498",
-  },
-  motu: {
-    title: "Murder of the Universe",
-    slug: "murder-of-the-universe",
-    releaseDate: "2017-06-23",
-    id: "45419e68-2701-4221-8c6d-ee29f66523de",
-    imageUrl: "https://f4.bcbits.com/img/a3406870115_16.jpg",
-    bandcampAlbumId: "1787442871",
-  },
-  sketches: {
-    title: "Sketches of Brunswick East",
-    slug: "sketches-of-brunswick-east",
-    releaseDate: "2017-08-18",
-    id: "a8fcce86-fab9-44d1-b3c2-1c6b1297ab2e",
-    imageUrl: "https://f4.bcbits.com/img/a4206546710_16.jpg",
-    bandcampAlbumId: "1255749163",
-  },
-  poly: {
-    title: "Polygondwanaland",
-    slug: "polygondwanaland",
-    releaseDate: "2017-11-17",
-    id: "03c0b2d8-626c-4c1c-b0f9-5c730477caab",
-    imageUrl: "https://f4.bcbits.com/img/a1550557789_16.jpg",
-    bandcampAlbumId: "1017207616",
-  },
-  gumbootSoup: {
-    title: "Gumboot Soup",
-    slug: "gumboot-soup",
-    releaseDate: "2017-12-31",
-    id: "08816b5c-5e1a-4aca-b07c-0a1b1008fabe",
-    imageUrl: "https://f4.bcbits.com/img/a1481325267_16.jpg",
-    bandcampAlbumId: "2728208210",
-  },
-  fff: {
-    title: "Fishing For Fishies",
-    slug: "fishing-for-fishies",
-    releaseDate: "2019-04-26",
-    id: "bc67b3f7-af1d-443f-9d28-1c99f6d4aa60",
-    imageUrl: "https://f4.bcbits.com/img/a0470322543_16.jpg",
-    bandcampAlbumId: "1507780262",
-  },
-  itrn: {
-    title: "Infest the Rats' Nest",
-    slug: "infest-the-rats-nest",
-    releaseDate: "2019-08-16",
-    id: "e889bd13-f73f-40fc-8ea9-17a6c3e9ad26",
-    imageUrl: "https://f4.bcbits.com/img/a0402916090_16.jpg",
-    bandcampAlbumId: "3925296262",
-  },
-  kg: {
-    title: "K.G.",
-    slug: "kg",
-    releaseDate: "2020-11-20",
-    id: "0c1f66d2-4ebc-499d-9bd7-c1ea34e4fbc0",
-    imageUrl: "https://f4.bcbits.com/img/a3546319741_16.jpg",
-    bandcampAlbumId: "927421026",
-  },
-  lw: {
-    title: "L.W.",
-    slug: "lw",
-    releaseDate: "2021-02-26",
-    id: "0b3e78ca-cfd3-4e74-8c47-5495271d0cc2",
-    imageUrl: "https://f4.bcbits.com/img/a2018491224_16.jpg",
-    bandcampAlbumId: "94095409",
-  },
-  bf3k: {
-    title: "Butterfly 3000",
-    slug: "butterfly-3000",
-    releaseDate: "2021-06-11",
-    id: "40fdce93-4438-4700-8721-9fade20341a0",
-    imageUrl: "https://f4.bcbits.com/img/a3009849752_16.jpg",
-    bandcampAlbumId: "3244133050",
-  },
-  og: {
-    title: "Omnium Gatherum",
-    slug: "omnium-gatherum",
-    releaseDate: "2022-04-22",
-    id: "560b0194-19ae-4bc9-aa24-0c2340596af7",
-    imageUrl: "https://f4.bcbits.com/img/a2638998972_16.jpg",
-    bandcampAlbumId: "129981996",
-  },
-  iceDeath: {
-    title: "Ice, Death, Planets, Lungs, Mushrooms and Lava",
-    slug: "ice-death-planets-lungs-mushrooms-and-lava",
-    releaseDate: "2022-07-10",
-    id: "46f925f1-dd7f-4966-8c02-3562e32f22b5",
-    imageUrl: "https://f4.bcbits.com/img/a3317842641_16.jpg",
-    bandcampAlbumId: "980855912",
-  },
-  laminatedDenim: {
-    title: "Laminated Denim",
-    slug: "laminate-denim",
-    releaseDate: "2022-10-12",
-    id: "7ceb3610-4aac-41cb-9a2f-4b6633dfad9b",
-    imageUrl: "https://f4.bcbits.com/img/a3665205369_16.jpg",
-    bandcampAlbumId: "177111450",
-  },
-  changes: {
-    title: "Changes",
-    slug: "changes",
-    releaseDate: "2022-10-28",
-    id: "957eae4b-eec8-45de-abcc-8902cda88fd3",
-    imageUrl: "https://f4.bcbits.com/img/a1890146474_16.jpg",
-    bandcampAlbumId: "2202545504",
-  },
-  petro: {
-    title:
-      "PetroDragonic Apocalypse; or, Dawn of Eternal Night: An Annihilation of Planet Earth and the Beginning of Merciless Damnation",
-    slug: "petro",
-    releaseDate: "2023-06-16",
-    id: "f43bb6f3-6b08-4687-a39e-a4f379325f93",
-    imageUrl: "https://f4.bcbits.com/img/a2805471381_16.jpg",
-    bandcampAlbumId: "3699259806",
-  },
-  tsc: {
-    title: "The Silver Cord",
-    slug: "the-silver-cord",
-    releaseDate: "2023-10-27",
-    id: "e212c176-b78f-4b0b-93a0-651ca99591e2",
-    imageUrl: "https://f4.bcbits.com/img/a0297097933_16.jpg",
-    bandcampAlbumId: "2151949833",
-  },
-  b741: {
-    title: "Flight b741",
-    slug: "flight-b741",
-    releaseDate: "2024-08-09",
-    id: "3e106926-3ea8-4742-87b4-7ea809799370",
-    imageUrl: "https://f4.bcbits.com/img/a3548545002_16.jpg",
-    bandcampAlbumId: "104852051",
-  },
-} as const satisfies Record<string, Omit<Album, "createdAt">>;
-export const allAlbums = Object.values(seedAlbums) as Album[];
-
-// =============================================================================
-// ================================= SONGS =====================================
-// =============================================================================
-
-// Ordered by album release date and position on album
-export const seedSongs = {
-  // ====================================================
-  // 12 Bar Bruise
-  elbow: {
-    id: "da1448b1-fec3-448b-9b6c-ed19b8c37a48",
-    albumPosition: 1,
-    slug: "elbow",
-    title: "Elbow",
-    albumId: seedAlbums.tbb.id,
-  },
-  muckraker: {
-    id: "d6095fe7-5f59-4b80-b885-af1e3b8e8984",
-    albumPosition: 2,
-    slug: "muckraker",
-    title: "Muckraker",
-    albumId: seedAlbums.tbb.id,
-  },
-  nein: {
-    id: "3e102507-1f85-416a-9047-62fdded23bf2",
-    albumPosition: 3,
-    slug: "nein",
-    title: "Nein",
-    albumId: seedAlbums.tbb.id,
-  },
-  twelveBarBruise: {
-    id: "b10f7d10-d460-45d8-a23d-380b071a63f1",
-    albumPosition: 4,
-    slug: "12-bar-bruise",
-    title: "12 Bar Bruise",
-    albumId: seedAlbums.tbb.id,
-  },
-  garageLiddiard: {
-    id: "e5c76308-5318-4736-9ccb-29c81c20673d",
-    albumPosition: 5,
-    slug: "garage-liddiard",
-    title: "Garage Liddiard",
-    albumId: seedAlbums.tbb.id,
-  },
-  samCherry: {
-    id: "d7630721-82d8-4acb-950a-e7b45d17a0c9",
-    albumPosition: 6,
-    slug: "sam-cherrys-last-shot",
-    title: "Sam Cherry's Last Shot",
-    albumId: seedAlbums.tbb.id,
-  },
-  highHopesLow: {
-    id: "ab8345f7-b1a5-4a9f-bd23-3e44ffd159b4",
-    albumPosition: 7,
-    slug: "high-hopes-low",
-    title: "High Hopes Low",
-    albumId: seedAlbums.tbb.id,
-  },
-  cutThroatBoogie: {
-    id: "b72be442-0339-42ea-be23-249c4207428e",
-    albumPosition: 8,
-    slug: "cut-throat-boogie",
-    title: "Cut Throat Boogie",
-    albumId: seedAlbums.tbb.id,
-  },
-  bloodyRipper: {
-    id: "70067e17-6ad4-4057-aeb2-4ea7cb614104",
-    albumPosition: 9,
-    slug: "bloody-ripper",
-    title: "Bloody Ripper",
-    albumId: seedAlbums.tbb.id,
-  },
-  uhOhICalledMum: {
-    id: "7e21c3f7-da5e-4890-846c-28765bf5d487",
-    albumPosition: 10,
-    slug: "uh-oh-i-called-mum",
-    title: "Up Oh I Called Mum",
-    albumId: seedAlbums.tbb.id,
-  },
-  seaOfTrees: {
-    id: "43d67e96-d504-43a1-bcc3-7f25ac648e81",
-    albumPosition: 11,
-    slug: "sea-of-trees",
-    title: "Sea of Trees",
-    albumId: seedAlbums.tbb.id,
-  },
-  footyFooty: {
-    id: "8dbb87e9-509b-41c2-ab97-4de6bec1551e",
-    albumPosition: 12,
-    slug: "footy-footy",
-    title: "Footy Footy",
-    albumId: seedAlbums.tbb.id,
-  },
-
-  // ====================================================
-  // Float Along - Fill Your Lungs
-  headOn: {
-    id: "556e6f5e-9099-4351-9c7e-0edb17e76b73",
-    albumPosition: 1,
-    slug: "head-on-pill",
-    title: "Head On/Pill",
-    albumId: seedAlbums.fafyl.id,
-  },
-  iAmNotAMan: {
-    id: "8e4ef0bb-acfe-4eb2-838b-1782271dea28",
-    albumPosition: 2,
-    slug: "i-am-not-a-man-unless-i-have-a-woman",
-    title: "I Am Not A Man Unless I Have A Woman",
-    albumId: seedAlbums.fafyl.id,
-  },
-  godIsCallingMeBackHome: {
-    id: "195794b2-40d5-4101-9aa7-1cb1725168b6",
-    albumPosition: 3,
-    slug: "god-is-calling-me-back-home",
-    title: "God Is Calling Me Back Home",
-    albumId: seedAlbums.fafyl.id,
-  },
-  thirtyPastSeven: {
-    id: "42455238-cdb5-4356-92db-688d2a4708f9",
-    albumPosition: 4,
-    slug: "30-past-7",
-    title: "30 Past 7",
-    albumId: seedAlbums.fafyl.id,
-  },
-  letMeMendThePast: {
-    id: "fa5d8f49-6cee-40f7-b4ea-e68de744035e",
-    albumPosition: 5,
-    slug: "let-me-mend-the-past",
-    title: "Let Me Mend the Past",
-    albumId: seedAlbums.fafyl.id,
-  },
-  mysteryJack: {
-    id: "87980647-e70a-456d-827a-210ca134ebce",
-    albumPosition: 6,
-    slug: "mystery-jack",
-    title: "Mystery Jack",
-    albumId: seedAlbums.fafyl.id,
-  },
-  popInMyStep: {
-    id: "0230ae45-7e08-4ec6-b585-162a03a0c830",
-    albumPosition: 7,
-    slug: "pop-in-my-step",
-    title: "Pop In My Step",
-    albumId: seedAlbums.fafyl.id,
-  },
-  fafyl: {
-    id: "4a0aac05-3060-4750-9eb0-fd6a058fbfc1",
-    albumPosition: 8,
-    slug: "float-along-fill-your-lungs",
-    title: "Float Along - Fill Your Lungs",
-    albumId: seedAlbums.fafyl.id,
-  },
-
-  // ====================================================
-  // Oddments
-  alludaMajaka: {
-    id: "001b0a76-cf88-4d4f-9f82-0a1b4feca97d",
-    albumPosition: 1,
-    slug: "alluda-majaka",
-    title: "Alluda Majaka",
-    albumId: seedAlbums.oddments.id,
-  },
-  stressin: {
-    id: "ec5f8e95-a9c8-4a79-b16f-5ac56be4c5ba",
-    albumPosition: 2,
-    slug: "stressin",
-    title: "Stressin'",
-    albumId: seedAlbums.oddments.id,
-  },
-  vegemite: {
-    id: "917d455e-421c-4099-8765-9fa8876501b6",
-    albumPosition: 3,
-    slug: "vegemite",
-    title: "Vegemite",
-    albumId: seedAlbums.oddments.id,
-  },
-  itsGotOld: {
-    id: "ede4ff9e-b9b8-464e-b9e1-3efe95a63cbf",
-    albumPosition: 4,
-    slug: "its-got-old",
-    title: "It's Got Old",
-    albumId: seedAlbums.oddments.id,
-  },
-  workThisTime: {
-    id: "460537ac-78a5-4358-80d4-e60a3aa78b93",
-    albumPosition: 5,
-    slug: "work-this-time",
-    title: "Work This Time",
-    albumId: seedAlbums.oddments.id,
-  },
-  ababcd: {
-    id: "933fb599-0e0a-4d40-b34c-fb4fd11be788",
-    albumPosition: 6,
-    slug: "ababcd",
-    title: "ABABCD",
-    albumId: seedAlbums.oddments.id,
-  },
-  sleepwalker: {
-    id: "bb4d4d97-9674-45b0-b5d6-95b22dbf934e",
-    albumPosition: 7,
-    slug: "sleepwalker",
-    title: "Sleepwalker",
-    albumId: seedAlbums.oddments.id,
-  },
-  hotWax: {
-    id: "57430f87-f009-4094-8c66-e08b8041bbc3",
-    albumPosition: 8,
-    slug: "hot-wax",
-    title: "Hot Wax",
-    albumId: seedAlbums.oddments.id,
-  },
-  crying: {
-    id: "cf529fbc-3639-44cf-b28c-a846c2195e33",
-    albumPosition: 9,
-    slug: "crying",
-    title: "Crying",
-    albumId: seedAlbums.oddments.id,
-  },
-  pipeDream: {
-    id: "7b17ec59-af83-4196-9b98-21d1686c6b52",
-    albumPosition: 10,
-    slug: "pipe-dream",
-    title: "Pipe-Dream",
-    albumId: seedAlbums.oddments.id,
-  },
-  homelessManInAdidas: {
-    id: "5d566e25-89e0-4133-bef0-825263c94d14",
-    albumPosition: 11,
-    slug: "homeless-man-in-adidas",
-    title: "Homeless Man in Adidas",
-    albumId: seedAlbums.oddments.id,
-  },
-  oddments: {
-    id: "22102ac6-f795-4e33-848e-3689bb45e49c",
-    albumPosition: 12,
-    slug: "oddments",
-    title: "Oddments",
-    albumId: seedAlbums.oddments.id,
-  },
-
-  // ====================================================
-  // I'm In Your Mind Fuzz
-  imInYourMind: {
-    id: "7b4d6326-bdb9-4753-9076-0630b5a95b32",
-    albumPosition: 1,
-    slug: "im-in-your-mind",
-    title: "I'm In Your Mind",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  imNotInYourMind: {
-    id: "06eba6d7-d32f-40f1-85ca-32e79cdbb627",
-    albumPosition: 2,
-    slug: "im-not-in-your-mind",
-    title: "I'm Not In Your Mind",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  cellophane: {
-    id: "5ada7965-bbef-44ac-8317-f00cf8fcc17d",
-    albumPosition: 3,
-    slug: "cellaphane",
-    title: "Cellaphane",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  imInYourMindFuzz: {
-    id: "770a35de-91c6-456b-8e5e-06f71cb08b4c",
-    albumPosition: 4,
-    slug: "im-in-your-mind-fuzz",
-    title: "I'm In Your Mind Fuzz",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  empty: {
-    id: "b1d72add-e97a-4c2a-922d-fb13d56df828",
-    albumPosition: 5,
-    slug: "empty",
-    title: "Empty",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  hotWater: {
-    id: "7a0ac425-58be-4373-8f14-78905db4a081",
-    albumPosition: 6,
-    slug: "hot-water",
-    title: "Hot Water",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  amIInHeaven: {
-    id: "14c20df5-ff64-40b7-b410-aa2aa8daab83",
-    albumPosition: 7,
-    slug: "am-i-in-heaven",
-    title: "Am I in Heaven?",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  slowJam1: {
-    id: "d8206c9b-096f-4c9b-a477-ec36ed88de72",
-    albumPosition: 8,
-    slug: "slow-jam-1",
-    title: "Slow Jam 1",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  satanSpeedsUp: {
-    id: "1b6fad18-50a7-4a52-93d2-d76d71d59142",
-    albumPosition: 9,
-    slug: "satan-speeds-up",
-    title: "Satan Speeds Up",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-  herAndI: {
-    id: "a0d4e6aa-e5d8-4402-82ef-27918958dafd",
-    albumPosition: 10,
-    slug: "her-and-i-slow-jam-2",
-    title: "Her and I (Slow Jam 2)",
-    albumId: seedAlbums.mindFuzz.id,
-  },
-
-  // ====================================================
-  // Quarters
-  theRiver: {
-    id: "5855d6aa-2dd1-4f82-a2d8-44b4e663eb87",
-    albumPosition: 1,
-    slug: "the-river",
-    title: "The River",
-    albumId: seedAlbums.quarters.id,
-  },
-  infiniteRise: {
-    id: "a0f65c26-32a9-4c54-a282-da281df11a9a",
-    albumPosition: 2,
-    slug: "infinte-rise",
-    title: "Infinite Rise",
-    albumId: seedAlbums.quarters.id,
-  },
-  godIsInTheRhythm: {
-    id: "e7143dd3-f080-4e63-943a-1c51e0458e0c",
-    albumPosition: 3,
-    slug: "god-is-in-the-rhythm",
-    title: "God Is In The Rhythm",
-    albumId: seedAlbums.quarters.id,
-  },
-  lonelySteelSheetFlyer: {
-    id: "d946c819-4bb8-4d3e-bc9a-07c6cd8ba410",
-    albumPosition: 4,
-    slug: "lonely-steel-sheet-flyer",
-    title: "Lonely Steel Sheet Flyer",
-    albumId: seedAlbums.quarters.id,
-  },
-
-  // ====================================================
-  // Paper Mâché Dream Balloon
-  sense: {
-    id: "d916bb15-6c5d-4d16-8224-4a4d30bd09b7",
-    albumPosition: 1,
-    slug: "sense",
-    title: "Sense",
-    albumId: seedAlbums.pmdb.id,
-  },
-  bone: {
-    id: "e07317d8-1778-40ab-b9f5-9fc554580041",
-    albumPosition: 2,
-    slug: "bone",
-    title: "Bone",
-    albumId: seedAlbums.pmdb.id,
-  },
-  dirt: {
-    id: "b0550f16-5560-4184-8fa7-d5c0cf624522",
-    albumPosition: 3,
-    slug: "dirt",
-    title: "Dirt",
-    albumId: seedAlbums.pmdb.id,
-  },
-  paperMacheDreamBalloon: {
-    id: "b3b3d82b-aa4c-4f57-b29b-883dd587d174",
-    albumPosition: 4,
-    slug: "paper-mache-dream-balloon",
-    title: "Paper Mâché Dream Balloon",
-    albumId: seedAlbums.pmdb.id,
-  },
-  trapdoor: {
-    id: "7e5a8135-fab6-4dce-b54e-b784d94a7258",
-    albumPosition: 5,
-    slug: "trapdoor",
-    title: "Trapdoor",
-    albumId: seedAlbums.pmdb.id,
-  },
-  coldCadaver: {
-    id: "1b7998e0-b979-4fd6-85ce-b5fada63e3df",
-    albumPosition: 6,
-    slug: "cold-cadaver",
-    title: "Cold Cadaver",
-    albumId: seedAlbums.pmdb.id,
-  },
-  theBitterBoogie: {
-    id: "c613a961-1a79-491b-a567-995e018ba74d",
-    albumPosition: 7,
-    slug: "the-bitter-boogie",
-    title: "The Bitter Boogie",
-    albumId: seedAlbums.pmdb.id,
-  },
-  ngri: {
-    id: "7f3df9c6-1263-440f-bfa4-0a59c40fdcc9",
-    albumPosition: 8,
-    slug: "ngri-bloodstain",
-    title: "N.G.R.I. (Bloodstain)",
-    albumId: seedAlbums.pmdb.id,
-  },
-  timeFate: {
-    id: "e47a83d3-6f33-43c3-ac6b-afc91f22e666",
-    albumPosition: 9,
-    slug: "time-fate",
-    title: "Time = Fate",
-    albumId: seedAlbums.pmdb.id,
-  },
-  timeMoney: {
-    id: "8538e9ce-a33b-4de9-85bf-97ae6f35b9df",
-    albumPosition: 10,
-    slug: "time-money",
-    title: "Time = $$$",
-    albumId: seedAlbums.pmdb.id,
-  },
-  mostOfWhatILike: {
-    id: "e15852a6-5f13-410e-95af-f85f3d698350",
-    albumPosition: 11,
-    slug: "most-of-what-i-like",
-    title: "Most Of What I Like",
-    albumId: seedAlbums.pmdb.id,
-  },
-  paperMache: {
-    id: "1d6351a3-0466-49cd-86fb-96d3db6d3e6c",
-    albumPosition: 12,
-    slug: "paper-mache",
-    title: "Paper Mâché",
-    albumId: seedAlbums.pmdb.id,
-  },
-
-  // ====================================================
-  // Nonagon Infinity
-  robotStop: {
-    id: "047f1192-9893-4776-882f-d76542a71943",
-    albumPosition: 1,
-    slug: "robot-stop",
-    title: "Robot Stop",
-    albumId: seedAlbums.nonagon.id,
-  },
-  bigFigWasp: {
-    id: "9c4c49aa-9a0b-40c1-9873-d30db334eb4a",
-    albumPosition: 2,
-    slug: "big-fig-wasp",
-    title: "Big Fig Wasp",
-    albumId: seedAlbums.nonagon.id,
-  },
-  gammaKnife: {
-    id: "ae61473d-3e8d-4b83-b48e-2025c9c1b9a0",
-    albumPosition: 3,
-    slug: "gamma-knife",
-    title: "Gamma Knife",
-    albumId: seedAlbums.nonagon.id,
-  },
-  peopleVultures: {
-    id: "8bce52d7-bb5e-43fa-8a20-e3660e206f7a",
-    albumPosition: 4,
-    slug: "people-vultures",
-    title: "People Vultures",
-    albumId: seedAlbums.nonagon.id,
-  },
-  mrBeat: {
-    id: "8ec206bb-f1e7-45ed-90a7-c05233fb6de6",
-    albumPosition: 5,
-    slug: "mr-beat",
-    title: "Mr. Beat",
-    albumId: seedAlbums.nonagon.id,
-  },
-  evilDeathRoll: {
-    id: "6eb610e1-1391-4ad3-ad61-5770fcfc02b8",
-    albumPosition: 6,
-    slug: "evil-death-roll",
-    title: "Evil Death Roll",
-    albumId: seedAlbums.nonagon.id,
-  },
-  invisibleFace: {
-    id: "214d2f6a-86f7-4825-b0e1-f4b7b5b2d298",
-    albumPosition: 7,
-    slug: "invisible-face",
-    title: "Invisible Face",
-    albumId: seedAlbums.nonagon.id,
-  },
-  wahWah: {
-    id: "f57a85a9-37bb-4a09-acbe-c9b595f03792",
-    albumPosition: 8,
-    slug: "wah-wah",
-    title: "Wah Wah",
-    albumId: seedAlbums.nonagon.id,
-  },
-  roadTrain: {
-    id: "d659a3f1-f66e-4583-9d52-655ea5441cb7",
-    albumPosition: 9,
-    slug: "road-train",
-    title: "Road Train",
-    albumId: seedAlbums.nonagon.id,
-  },
-
-  // ====================================================
-  // Flying Microtonal Banana
-  rattlesnake: {
-    id: "4dc62f80-ad92-4f8d-8e66-bbe29d3e41a1",
-    albumPosition: 1,
-    slug: "rattlesnake",
-    title: "Rattlesnake",
-    albumId: seedAlbums.fmb.id,
-  },
-  melting: {
-    id: "ad82ad12-8d25-4738-b55a-2620d9cb0e7a",
-    albumPosition: 2,
-    slug: "melting",
-    title: "Melting",
-    albumId: seedAlbums.fmb.id,
-  },
-  openWater: {
-    id: "b2906883-cc2a-4bd6-96c4-6c728b60dd54",
-    albumPosition: 3,
-    slug: "open-water",
-    title: "Open Water",
-    albumId: seedAlbums.fmb.id,
-  },
-  sleepDrifter: {
-    id: "f65410dd-18bf-4bd0-8323-2f48ca376ae4",
-    albumPosition: 4,
-    slug: "sleep-drifter",
-    title: "Sleep Drifter",
-    albumId: seedAlbums.fmb.id,
-  },
-  billabongValley: {
-    id: "26a34912-4213-4d6b-b9db-9c8098cc636b",
-    albumPosition: 5,
-    slug: "billabong-valley",
-    title: "Billabong Valley",
-    albumId: seedAlbums.fmb.id,
-  },
-  anoxia: {
-    id: "fe1b296d-e29c-4876-b4c4-57c6b02c6510",
-    albumPosition: 6,
-    slug: "anoxia",
-    title: "Anoxia",
-    albumId: seedAlbums.fmb.id,
-  },
-  doomCity: {
-    id: "e27000c7-6d64-4103-8174-e0abfab2f1e6",
-    albumPosition: 7,
-    slug: "doom-city",
-    title: "Doom City",
-    albumId: seedAlbums.fmb.id,
-  },
-  nuclearFusion: {
-    id: "cbebf078-8eeb-40a7-b893-53b70877db0b",
-    albumPosition: 8,
-    slug: "nuclear-fusion",
-    title: "Nuclear Fusion",
-    albumId: seedAlbums.fmb.id,
-  },
-  flyingMicrotonalBanana: {
-    id: "5a5887aa-fc91-4fbe-bd52-e3405accc12e",
-    albumPosition: 9,
-    slug: "flyin-microtonal-banana",
-    title: "Flying Microtonal Banana",
-    albumId: seedAlbums.fmb.id,
-  },
-
-  // ====================================================
-  // Murder of the Universe
-  aNewWorld: {
-    id: "ec659f44-f428-4175-b190-40458c24fee7",
-    albumPosition: 1,
-    slug: "a-new-world",
-    title: "A New World",
-    albumId: seedAlbums.motu.id,
-  },
-  alteredBeastI: {
-    id: "da09da99-231c-444e-ab5b-752656705d0f",
-    albumPosition: 2,
-    slug: "altered-beast-i",
-    title: "Altered Beast I",
-    albumId: seedAlbums.motu.id,
-  },
-  alterMeI: {
-    id: "b876df42-8446-4a1c-9461-38f5d39b8829",
-    albumPosition: 3,
-    slug: "alter-me-i",
-    title: "Alter Me I",
-    albumId: seedAlbums.motu.id,
-  },
-  alteredBeastII: {
-    id: "b38abf6c-e6c3-4fae-b861-ebf4e1900738",
-    albumPosition: 4,
-    slug: "altered-beast-ii",
-    title: "Altered Beast II",
-    albumId: seedAlbums.motu.id,
-  },
-  alterMeII: {
-    id: "cc5c347e-1281-4fb3-9759-17a3b0c9cc0e",
-    albumPosition: 5,
-    slug: "alter-me-ii",
-    title: "Alter Me II",
-    albumId: seedAlbums.motu.id,
-  },
-  alteredBeastIII: {
-    id: "05691ee6-b19d-4dad-a7b7-5fd8eda5e783",
-    albumPosition: 6,
-    slug: "altered-beast-iii",
-    title: "Altered Beast III",
-    albumId: seedAlbums.motu.id,
-  },
-  alteredMeIII: {
-    id: "50a1717d-6262-4109-8e28-11635bf615a1",
-    albumPosition: 7,
-    slug: "altered-me-iii",
-    title: "Altered Me III",
-    albumId: seedAlbums.motu.id,
-  },
-  alteredBeastIV: {
-    id: "25c5876c-5d49-41f0-b227-5336dd9383ad",
-    albumPosition: 8,
-    slug: "altered-beast-iv",
-    title: "Altered Beast IV",
-    albumId: seedAlbums.motu.id,
-  },
-  lifeDeath: {
-    id: "a16e5705-3e25-435a-8027-075475427651",
-    albumPosition: 9,
-    slug: "life-death",
-    title: "Life / Death",
-    albumId: seedAlbums.motu.id,
-  },
-  someContext: {
-    id: "3106f462-3e2c-4e85-a04b-f9287355eed1",
-    albumPosition: 10,
-    slug: "some-context",
-    title: "Some Context",
-    albumId: seedAlbums.motu.id,
-  },
-  theReticentRaconteur: {
-    id: "262f0d2c-ac38-4094-97d4-f5832f448ca4",
-    albumPosition: 11,
-    slug: "the-reticent-raconteur",
-    title: "The Reticent Raconteur",
-    albumId: seedAlbums.motu.id,
-  },
-  lordOfLightning: {
-    id: "3dea7dc9-0738-4803-80cc-3dadd872a867",
-    albumPosition: 12,
-    slug: "lord-of-lightning",
-    title: "Lord of Lightning",
-    albumId: seedAlbums.motu.id,
-  },
-  theBalrog: {
-    id: "ff60523a-636a-412d-a27f-7a64c649bada",
-    albumPosition: 13,
-    slug: "the-balrog",
-    title: "The Balrog",
-    albumId: seedAlbums.motu.id,
-  },
-  theFloatingFire: {
-    id: "4c066743-8c66-476b-a8df-088a694bb80a",
-    albumPosition: 14,
-    slug: "the-floating-fire",
-    title: "The Floating Fire",
-    albumId: seedAlbums.motu.id,
-  },
-  theAcridCorpse: {
-    id: "6c3079b0-750e-496c-8aec-2c240997c5aa",
-    albumPosition: 15,
-    slug: "the-acrid-corpse",
-    title: "The Acrid Corpse",
-    albumId: seedAlbums.motu.id,
-  },
-  welcomeToAnAlteredFuture: {
-    id: "2e9595e8-0c29-422b-bcde-4d141746bde7",
-    albumPosition: 16,
-    slug: "welcome-to-an-altered-future",
-    title: "Welcome to an Altered Future",
-    albumId: seedAlbums.motu.id,
-  },
-  digitalBlack: {
-    id: "e03438c5-16d6-4855-8106-1eca75274c2b",
-    albumPosition: 17,
-    slug: "digital-black",
-    title: "Digital Black",
-    albumId: seedAlbums.motu.id,
-  },
-  hanTyumiTheConfusedCyborg: {
-    id: "00aa6690-b64c-4511-a747-c864f047efdc",
-    albumPosition: 18,
-    slug: "han-tyumi-the-confused-cyborg",
-    title: "Han-Tyumi, the Confused Cyborg",
-    albumId: seedAlbums.motu.id,
-  },
-  soyProtienMuntMachine: {
-    id: "65f73606-3059-4a38-8378-93c0b2a265ab",
-    albumPosition: 19,
-    slug: "soy-protein-munt-machine",
-    title: "Soy Protein Munt Machine",
-    albumId: seedAlbums.motu.id,
-  },
-  vomitCoffin: {
-    id: "2453de9d-2431-4174-8561-94f5c2f08d21",
-    albumPosition: 20,
-    slug: "vomit-coffin",
-    title: "Vomit Coffin",
-    albumId: seedAlbums.motu.id,
-  },
-  motu: {
-    id: "3bed1339-307c-48eb-b17e-ba582f0c00b6",
-    albumPosition: 21,
-    slug: "murder-of-the-universe",
-    title: "Murder of the Universe",
-    albumId: seedAlbums.motu.id,
-  },
-
-  // ====================================================
-  // Sketches of Brunswick East
-  sketchesOfBruswickEastI: {
-    id: "96ca5a82-6be2-489f-b875-bc9447e98be5",
-    albumPosition: 1,
-    slug: "sketches-of-brunswick-east-i",
-    title: "Sketches Of Brunswick East I",
-    albumId: seedAlbums.sketches.id,
-  },
-  countdown: {
-    id: "bc316fd6-c451-4c7d-b984-ab9e4b39b035",
-    albumPosition: 2,
-    slug: "countdown",
-    title: "Countdown",
-    albumId: seedAlbums.sketches.id,
-  },
-  dDay: {
-    id: "acd643e5-5293-42d0-952d-be7286e660b2",
-    albumPosition: 3,
-    slug: "d-day",
-    title: "D-Day",
-    albumId: seedAlbums.sketches.id,
-  },
-  tezeta: {
-    id: "aaecdc78-f296-4aaf-9427-f30ee33c29e4",
-    albumPosition: 4,
-    slug: "tezeta",
-    title: "Tezeta",
-    albumId: seedAlbums.sketches.id,
-  },
-  cranesPlanesMigraines: {
-    id: "a08f11ba-03a8-429c-b675-a71f7a11f1db",
-    albumPosition: 5,
-    slug: "cranes-planes-migraines",
-    title: "Cranes, Planes, Migraines",
-    albumId: seedAlbums.sketches.id,
-  },
-  theSpiderAndMe: {
-    id: "66a2c110-a364-49d8-8edc-c0eee1f9bc0f",
-    albumPosition: 6,
-    slug: "the-spider-and-me",
-    title: "The Spider And Me",
-    albumId: seedAlbums.sketches.id,
-  },
-  sketchesOfBrunswickEastII: {
-    id: "9f9d10ca-483e-4dac-b91e-d3aeaf91e61d",
-    albumPosition: 7,
-    slug: "sketches-of-brunswick-east-ii",
-    title: "Sketches Of Brunswick East II",
-    albumId: seedAlbums.sketches.id,
-  },
-  duskToDawnOnLygonStreet: {
-    id: "7a6ab1f1-5553-4164-be29-6b82b50c3533",
-    albumPosition: 8,
-    slug: "dusk-to-dawn-on-lygon-street",
-    title: "Dusk To Dawn On Lygon Street",
-    albumId: seedAlbums.sketches.id,
-  },
-  theBook: {
-    id: "d7f653a1-f774-4462-8ad6-dde97a0bce4f",
-    albumPosition: 9,
-    slug: "the-book",
-    title: "The Book",
-    albumId: seedAlbums.sketches.id,
-  },
-  aJourneyToShell: {
-    id: "2e21628f-0418-45ec-88f3-43d7561d1346",
-    albumPosition: 10,
-    slug: "a-journey-to-shell",
-    title: "A Journey To (S)hell",
-    albumId: seedAlbums.sketches.id,
-  },
-  rollingStoned: {
-    id: "c8c5ac74-452d-40a7-8fde-e4d4caa63ae0",
-    albumPosition: 11,
-    slug: "rolling-stoned",
-    title: "Rolling Stoned",
-    albumId: seedAlbums.sketches.id,
-  },
-  silhouette: {
-    id: "84be6b7d-8ef2-4a77-ad41-f6c8ab98e682",
-    albumPosition: 12,
-    slug: "you-can-be-your-silhouette",
-    title: "You Can Be Your Silhouette",
-    albumId: seedAlbums.sketches.id,
-  },
-  sketchesOfBrunswickEastIII: {
-    id: "79209906-5cc6-4b30-b8b2-e6664e4607cf",
-    albumPosition: 13,
-    slug: "sketches-of-brunswick-east-iii",
-    title: "Sketches Of Brunswick East III",
-    albumId: seedAlbums.sketches.id,
-  },
-
-  // ====================================================
-  // Polygondwanaland
-  crumblingCastle: {
-    id: "46b73130-6997-466f-b570-4d4528913f9c",
-    albumPosition: 1,
-    slug: "crumbling-castle",
-    title: "Crumbling Castle",
-    albumId: seedAlbums.poly.id,
-  },
-  polygondwanaland: {
-    id: "c59cbe5a-0e9f-4421-88c8-5e9ad708782a",
-    albumPosition: 2,
-    slug: "polygondwanaland",
-    title: "Polygondwanaland",
-    albumId: seedAlbums.poly.id,
-  },
-  theCastleInTheAir: {
-    id: "e9ae5d00-4f87-43ff-9a28-21c2ad82604e",
-    albumPosition: 3,
-    slug: "the-castle-in-the-air",
-    title: "The Castle In The Air",
-    albumId: seedAlbums.poly.id,
-  },
-  desertedDunesWelcomeWearyFeet: {
-    id: "6e94266e-2cdc-489e-9c10-82aee350aea0",
-    albumPosition: 4,
-    slug: "deserted-dunes-welcome-weary-feet",
-    title: "Deserted Dunes Welcome Weary Feet",
-    albumId: seedAlbums.poly.id,
-  },
-  innerCell: {
-    id: "b9fd1057-858e-4c27-bbea-eb6a3d0e00db",
-    albumPosition: 5,
-    slug: "inner-cell",
-    title: "Inner Cell",
-    albumId: seedAlbums.poly.id,
-  },
-  loyalty: {
-    id: "5aceb4cc-9b8f-4e1a-8d01-8d0bf129ec1f",
-    albumPosition: 6,
-    slug: "loyalty",
-    title: "Loaylty",
-    albumId: seedAlbums.poly.id,
-  },
-  horology: {
-    id: "005d7bee-2d21-4d4f-8535-8d7c5678f041",
-    albumPosition: 7,
-    slug: "horology",
-    title: "Horology",
-    albumId: seedAlbums.poly.id,
-  },
-  tetrachromacy: {
-    id: "f6b0a28f-fa34-49c2-a597-e637ee13852e",
-    albumPosition: 8,
-    slug: "tetrachromacy",
-    title: "Tetrachromacy",
-    albumId: seedAlbums.poly.id,
-  },
-  searching: {
-    id: "94d43344-d0f3-4e8e-b45d-fb867039357f",
-    albumPosition: 9,
-    slug: "searching",
-    title: "Searching...",
-    albumId: seedAlbums.poly.id,
-  },
-  fourthColor: {
-    id: "cbf2a357-406a-46df-b6cb-639b06282cef",
-    albumPosition: 10,
-    slug: "the-fourth-color",
-    title: "The Fourth Color",
-    albumId: seedAlbums.poly.id,
-  },
-
-  // ====================================================
-  // Gumboot Soup
-  beginnersLuck: {
-    id: "300516d1-2a31-4dc1-ae51-b7544f5fb698",
-    albumPosition: 1,
-    slug: "beginners-luck",
-    title: "Beginner's Luck",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  greenhouseHeatDeath: {
-    id: "6fc44241-d6b0-44b7-9eba-bdd4c08430dc",
-    albumPosition: 2,
-    slug: "greenhouse-heat-death",
-    title: "Greenhouse Heat Death",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  barefootDesert: {
-    id: "3596353d-b9fb-4b9f-a382-618320cfbfcc",
-    albumPosition: 3,
-    slug: "barefoot-desert",
-    title: "Barefoot Desert",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  muddyWater: {
-    id: "37476f9c-0e46-4ba6-9fb6-727cb904f567",
-    albumPosition: 4,
-    slug: "muddy-water",
-    title: "Muddy Water",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  superposition: {
-    id: "30f1491f-7b7e-4981-95c7-af82b64001d4",
-    albumPosition: 5,
-    slug: "superposition",
-    title: "Superposition",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  downTheSink: {
-    id: "f3df8a93-be8f-40e0-ab26-a9960328d5a5",
-    albumPosition: 6,
-    slug: "down-the-sink",
-    title: "Down The Sink",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  theGreatChainOfBeing: {
-    id: "44719352-8a93-4462-a4b5-afac9abad8ba",
-    albumPosition: 7,
-    slug: "the-great-chain-of-being",
-    title: "The Great Chain Of Being",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  theLastOasis: {
-    id: "0a23a3d0-014c-4fad-9f46-6e37af395a32",
-    albumPosition: 8,
-    slug: "the-last-oasis",
-    title: "The Last Oasis",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  allIsKnown: {
-    id: "1e83ee0c-2f9c-4013-8eb1-00f7ff696127",
-    albumPosition: 9,
-    slug: "all-is-known",
-    title: "All Is Known",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  imSleepinIn: {
-    id: "bd5c1d16-420e-4770-8fd1-cd2e3e2dd10f",
-    albumPosition: 10,
-    slug: "im-sleepin-in",
-    title: "I'm Sleepin' In",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-  theWheel: {
-    id: "20fc1b2b-6838-422a-bf6f-9edff1b74a2f",
-    albumPosition: 11,
-    slug: "the-wheel",
-    title: "The Wheel",
-    albumId: seedAlbums.gumbootSoup.id,
-  },
-
-  // ====================================================
-  // Fishing For Fishies
-  fishingForFishies: {
-    id: "7370f606-49b2-40eb-a9eb-58b84ab5446b",
-    albumPosition: 1,
-    slug: "fishing-for-fishies",
-    title: "Fishing For Fishies",
-    albumId: seedAlbums.fff.id,
-  },
-  boogiemanSam: {
-    id: "a9a6572e-e1a3-4557-ab1e-388fc4db3b49",
-    albumPosition: 2,
-    slug: "booogieman-sam",
-    title: "Boogieman Sam",
-    albumId: seedAlbums.fff.id,
-  },
-  theBirdSong: {
-    id: "55bd5fb5-0a33-4da8-bf48-b220aa816714",
-    albumPosition: 3,
-    slug: "the-bird-song",
-    title: "The Bird Song",
-    albumId: seedAlbums.fff.id,
-  },
-  plasticBoogie: {
-    id: "f49713e5-5ddd-4142-b2ef-22dbf7dbb9d7",
-    albumPosition: 4,
-    slug: "plastic-boogie",
-    title: "Plastic Boogie",
-    albumId: seedAlbums.fff.id,
-  },
-  theCruelMillenial: {
-    id: "f9551ec5-cfaf-45ea-840d-8832f60d10bb",
-    albumPosition: 5,
-    slug: "the-cruel-millenial",
-    title: "The Cruel Millenial",
-    albumId: seedAlbums.fff.id,
-  },
-  realsNotReal: {
-    id: "beadf77a-678c-4c88-b165-6e5850bfaa68",
-    albumPosition: 6,
-    slug: "reals-not-real",
-    title: "Real's Not Real",
-    albumId: seedAlbums.fff.id,
-  },
-  thisThing: {
-    id: "f4a714ed-0de5-43d3-a92c-14cf1c201c15",
-    albumPosition: 7,
-    slug: "this-thing",
-    title: "This Thing",
-    albumId: seedAlbums.fff.id,
-  },
-  acarine: {
-    id: "82088978-5bf8-48d1-8c01-ffb2950a9468",
-    albumPosition: 8,
-    slug: "acarine",
-    title: "Acarine",
-    albumId: seedAlbums.fff.id,
-  },
-  cyboogie: {
-    id: "694877de-c6ea-4774-acad-d6cab1e1e23c",
-    albumPosition: 9,
-    slug: "cyboogie",
-    title: "Cyboogie",
-    albumId: seedAlbums.fff.id,
-  },
-
-  // ====================================================
-  // Infest the Rats' Nest
-  planetB: {
-    id: "29fd0103-5c22-4210-9b19-4b59866ff3ae",
-    albumPosition: 1,
-    slug: "planet-b",
-    title: "Planet B",
-    albumId: seedAlbums.itrn.id,
-  },
-  marsForTheRich: {
-    id: "ef3d92e2-207d-4456-aff0-5399294c474d",
-    albumPosition: 2,
-    slug: "mars-for-the-rich",
-    title: "March for the Rich",
-    albumId: seedAlbums.itrn.id,
-  },
-  organFarmer: {
-    id: "c05da847-aee7-4354-815d-cbca6b8617b3",
-    albumPosition: 3,
-    slug: "organ-farmer",
-    title: "Organ Farmer",
-    albumId: seedAlbums.itrn.id,
-  },
-  superbug: {
-    id: "a03b54e1-154c-498b-87c4-ac5513352ded",
-    albumPosition: 4,
-    slug: "superbug",
-    title: "Superbug",
-    albumId: seedAlbums.itrn.id,
-  },
-  venusian1: {
-    id: "3fcc4acd-aab0-4bbe-81b2-59f4150b24da",
-    albumPosition: 5,
-    slug: "venusian-1",
-    title: "Venusian 1",
-    albumId: seedAlbums.itrn.id,
-  },
-  perihelion: {
-    id: "620be584-d629-4a52-8c71-f2e612a6cd06",
-    albumPosition: 6,
-    slug: "perihelion",
-    title: "Perihelion",
-    albumId: seedAlbums.itrn.id,
-  },
-  venusian2: {
-    id: "81617b99-9ff5-43d7-9730-e96cb34e97c3",
-    albumPosition: 7,
-    slug: "venusian-2",
-    title: "Venusion 2",
-    albumId: seedAlbums.itrn.id,
-  },
-  selfImmolate: {
-    id: "c6ef1672-f434-4257-a329-08a67a37686f",
-    albumPosition: 8,
-    slug: "self-immolate",
-    title: "Self-Immolate",
-    albumId: seedAlbums.itrn.id,
-  },
-  hell: {
-    id: "babd0c51-56c4-4b07-91ce-c734e797e876",
-    albumPosition: 9,
-    slug: "hell",
-    title: "Hell",
-    albumId: seedAlbums.itrn.id,
-  },
-
-  // ====================================================
-  // K.G.
-  kglwIntro: {
-    id: "75d9a982-57e2-4659-bdc3-4c4d6ad35ee8",
-    albumPosition: 1,
-    slug: "kglw-intro",
-    title: "K.G.L.W.",
-    albumId: seedAlbums.kg.id,
-  },
-  automation: {
-    id: "c4d3ec3e-c5df-469b-a213-2f18776f4574",
-    albumPosition: 2,
-    slug: "automation",
-    title: "Automation",
-    albumId: seedAlbums.kg.id,
-  },
-  minimumBrainSize: {
-    id: "d327fdcc-dae6-4b2e-afed-70279d18ab61",
-    albumPosition: 3,
-    slug: "minimum-brain-size",
-    title: "Minimum Brain Size",
-    albumId: seedAlbums.kg.id,
-  },
-  strawsInTheWind: {
-    id: "962270a4-42d6-4eab-bae9-c6b7ff3b2750",
-    albumPosition: 4,
-    slug: "straws-in-the-wind",
-    title: "Straws In The Wind",
-    albumId: seedAlbums.kg.id,
-  },
-  someOfUs: {
-    id: "26638384-de65-41ec-acb5-cf6770bb885d",
-    albumPosition: 5,
-    slug: "some-of-us",
-    title: "Some Of Us",
-    albumId: seedAlbums.kg.id,
-  },
-  ontology: {
-    id: "f858d07d-dae5-4c5b-a379-7a1b17723a6f",
-    albumPosition: 6,
-    slug: "ontology",
-    title: "Ontology",
-    albumId: seedAlbums.kg.id,
-  },
-  intrasport: {
-    id: "027fbc3c-620c-41f0-b734-f3d00262b41a",
-    albumPosition: 7,
-    slug: "intrasport",
-    title: "Intrasport",
-    albumId: seedAlbums.kg.id,
-  },
-  oddlife: {
-    id: "7594b803-623d-4893-8cc4-192faa4d914d",
-    albumPosition: 8,
-    slug: "oddlife",
-    title: "Oddlife",
-    albumId: seedAlbums.kg.id,
-  },
-  honey: {
-    id: "63f563bc-5d84-493b-8e09-d39745d59597",
-    albumPosition: 9,
-    slug: "honey",
-    title: "Honey",
-    albumId: seedAlbums.kg.id,
-  },
-  theHungryWolfOfFate: {
-    id: "010c8727-06d1-4f5c-aecd-7753717fb824",
-    albumPosition: 10,
-    slug: "the-hungry-wolf-of-fate",
-    title: "The Hungry Wolf Of Fate",
-    albumId: seedAlbums.kg.id,
-  },
-
-  // ====================================================
-  // L.W.
-  ifNotNowThenWhen: {
-    id: "af273f1e-0bd4-4eb7-a3d3-17b0544e165b",
-    albumPosition: 1,
-    slug: "if-not-now-then-when",
-    title: "If Not Now, Then When?",
-    albumId: seedAlbums.lw.id,
-  },
-  one: {
-    id: "7fa2f2a2-ee65-408b-960e-c6966424314b",
-    albumPosition: 2,
-    slug: "one",
-    title: "O.N.E.",
-    albumId: seedAlbums.lw.id,
-  },
-  pleura: {
-    id: "d6804751-ab0a-4faf-b4f1-75de7e97054a",
-    albumPosition: 3,
-    slug: "pleura",
-    title: "Pleura",
-    albumId: seedAlbums.lw.id,
-  },
-  supremeAscendency: {
-    id: "9a611b8d-2feb-49ff-9411-3ba8bdca3766",
-    albumPosition: 4,
-    slug: "supreme-ascendency",
-    title: "Supreme Ascendency",
-    albumId: seedAlbums.lw.id,
-  },
-  staticElectricity: {
-    id: "61cecc4c-4582-4245-aaba-a34ae80f5e3b",
-    albumPosition: 5,
-    slug: "static-electricity",
-    title: "Static Electricity",
-    albumId: seedAlbums.lw.id,
-  },
-  eastWestLink: {
-    id: "405a4bfb-60b2-49c2-a311-3f1e6c9f1d7d",
-    albumPosition: 6,
-    slug: "east-west-link",
-    title: "East West Link",
-    albumId: seedAlbums.lw.id,
-  },
-  ataraxia: {
-    id: "b374aaa5-c516-4c21-83d7-a70688427282",
-    albumPosition: 7,
-    slug: "ataraxia",
-    title: "Ataraxia",
-    albumId: seedAlbums.lw.id,
-  },
-  seeMe: {
-    id: "a8441f3d-9238-4ed5-95c6-a675b6f489e6",
-    albumPosition: 8,
-    slug: "see-me",
-    title: "See Me",
-    albumId: seedAlbums.lw.id,
-  },
-  kglwOutro: {
-    id: "d052487f-b799-4a93-8720-191af54c9cac",
-    albumPosition: 9,
-    slug: "kglw-outro",
-    title: "K.G.L.W.",
-    albumId: seedAlbums.lw.id,
-  },
-
-  // ====================================================
-  // Butterfly 3000
-  yours: {
-    id: "c3328c2a-6529-4406-81b2-db0fc92180aa",
-    albumPosition: 1,
-    slug: "yours",
-    title: "Yours",
-    albumId: seedAlbums.bf3k.id,
-  },
-  shanghai: {
-    id: "f927bc2e-d505-4cd6-8357-f3d73ded4403",
-    albumPosition: 2,
-    slug: "shanghai",
-    title: "Shanghai",
-    albumId: seedAlbums.bf3k.id,
-  },
-  dreams: {
-    id: "85b5ae24-e0ee-45c6-9c52-86050f32f072",
-    albumPosition: 3,
-    slug: "dreams",
-    title: "Dreams",
-    albumId: seedAlbums.bf3k.id,
-  },
-  blueMorpho: {
-    id: "4d08c17a-ebd2-4a64-a595-2535383518bb",
-    albumPosition: 4,
-    slug: "blue-morpho",
-    title: "Blue Morpho",
-    albumId: seedAlbums.bf3k.id,
-  },
-  interiorPeople: {
-    id: "96f2555a-796d-4dd1-8bfa-52c007a55190",
-    albumPosition: 5,
-    slug: "interior-people",
-    title: "Interior People",
-    albumId: seedAlbums.bf3k.id,
-  },
-  catchingSmoke: {
-    id: "39b27387-2b46-435d-b97f-93948516a0c4",
-    albumPosition: 6,
-    slug: "catching-smoke",
-    title: "Catching Smoke",
-    albumId: seedAlbums.bf3k.id,
-  },
-  twoOTwoKillerYear: {
-    id: "caf3b82c-6cbc-4033-87d0-0466c6165ae0",
-    albumPosition: 7,
-    slug: "2-02-killer-year",
-    title: "2.02 Killer Year",
-    albumId: seedAlbums.bf3k.id,
-  },
-  blackHotSoup: {
-    id: "6032d411-9fbc-4fbd-af5a-d04347a07fc3",
-    albumPosition: 8,
-    slug: "black-hot-soup",
-    title: "Black Hot Soup",
-    albumId: seedAlbums.bf3k.id,
-  },
-  yaLove: {
-    id: "8602f786-a7fe-4b4d-87ad-b73e2b667ba8",
-    albumPosition: 9,
-    slug: "ya-love",
-    title: "Ya Love",
-    albumId: seedAlbums.bf3k.id,
-  },
-  butterfly3000: {
-    id: "01db9af7-d4b9-4ae1-931b-703ca6d6f40a",
-    albumPosition: 10,
-    slug: "butterfly-3000",
-    title: "Butterfly 3000",
-    albumId: seedAlbums.bf3k.id,
-  },
-
-  // ====================================================
-  // Omnium Gatherum
-  theDrippingTap: {
-    id: "ae4a1559-f995-4ba0-bb49-6c1e34e196b3",
-    albumPosition: 1,
-    slug: "the-dripping-tap",
-    title: "The Dripping Tap",
-    albumId: seedAlbums.og.id,
-  },
-  magentaMountain: {
-    id: "9100f1b2-0e74-40bb-a945-0f59b54fff6d",
-    albumPosition: 2,
-    slug: "magenta-mountain",
-    title: "Magenta Mountain",
-    albumId: seedAlbums.og.id,
-  },
-  kepler22b: {
-    id: "82c104ae-76ab-4bb3-a9cd-3602c2346831",
-    albumPosition: 3,
-    slug: "kepler-22b",
-    title: "Kepler-22b",
-    albumId: seedAlbums.og.id,
-  },
-  gaia: {
-    id: "96ec69a2-701b-4eb6-85f2-5631c24f81f8",
-    albumPosition: 4,
-    slug: "gaia",
-    title: "Gaia",
-    albumId: seedAlbums.og.id,
-  },
-  ambergris: {
-    id: "1c34fd6f-6389-48ed-a482-a8f5ed51ec3d",
-    albumPosition: 5,
-    slug: "ambergris",
-    title: "Ambergris",
-    albumId: seedAlbums.og.id,
-  },
-  sadieSorceress: {
-    id: "076078ab-dd1b-469b-ac14-680c497cf503",
-    albumPosition: 6,
-    slug: "sadie-sorceress",
-    title: "Sadie Sorceress",
-    albumId: seedAlbums.og.id,
-  },
-  evilestMan: {
-    id: "d0e742d2-e852-46ba-897d-a6d5497820ad",
-    albumPosition: 7,
-    slug: "evilest-man",
-    title: "Evilest Man",
-    albumId: seedAlbums.og.id,
-  },
-  theGardenGoblin: {
-    id: "cee2fd24-6e75-4db7-8580-171bd4992767",
-    albumPosition: 8,
-    slug: "the-garden-goblin",
-    title: "The Garden Goblin",
-    albumId: seedAlbums.og.id,
-  },
-  blameItOnTheWeather: {
-    id: "68b520bd-0c75-4709-91a5-07feaf1b57dd",
-    albumPosition: 9,
-    slug: "blame-it-on-the-weather",
-    title: "Blame It On The Weather",
-    albumId: seedAlbums.og.id,
-  },
-  persistence: {
-    id: "c44d1887-b13e-487b-b545-0b13dced2fa0",
-    albumPosition: 10,
-    slug: "persistance",
-    title: "Persistence",
-    albumId: seedAlbums.og.id,
-  },
-  theGrimReaper: {
-    id: "99018b73-3191-49d3-a792-de9ea16b4a0e",
-    albumPosition: 11,
-    slug: "the-grim-reaper",
-    title: "The Grim Reaper",
-    albumId: seedAlbums.og.id,
-  },
-  presumptuous: {
-    id: "aff6b78b-28a5-4993-b538-ad203b5fba86",
-    albumPosition: 12,
-    slug: "presumptuous",
-    title: "Presumptuous",
-    albumId: seedAlbums.og.id,
-  },
-  predatorX: {
-    id: "0778fb7e-f1f1-4ff9-8a0c-0e3833f4f120",
-    albumPosition: 13,
-    slug: "predator-x",
-    title: "Predator X",
-    albumId: seedAlbums.og.id,
-  },
-  redSmoke: {
-    id: "9fbc3894-e2ee-42df-8fc1-8ec331ac9cd2",
-    albumPosition: 14,
-    slug: "red-smoke",
-    title: "Red Smoke",
-    albumId: seedAlbums.og.id,
-  },
-  candles: {
-    id: "e8f3aa5c-c621-484a-a95e-cad79dafdcb9",
-    albumPosition: 15,
-    slug: "candles",
-    title: "Candles",
-    albumId: seedAlbums.og.id,
-  },
-  theFuneral: {
-    id: "2ea76ba7-25e4-430d-b0aa-9ca37eae36f4",
-    albumPosition: 16,
-    slug: "the-funeral",
-    title: "The Funeral",
-    albumId: seedAlbums.og.id,
-  },
-
-  // ====================================================
-  // Ice, Death, Planets, Lungs, Mushrooms and Lava
-  mycelium: {
-    id: "484a071d-8043-483a-a85a-8b8183f73326",
-    albumPosition: 1,
-    slug: "mycelium",
-    title: "Mycelium",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  iceV: {
-    id: "ae36e961-6db2-4c3b-8f8d-d08ecb671543",
-    albumPosition: 2,
-    slug: "ice-v",
-    title: "Ice V",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  magma: {
-    id: "717f22ee-9458-4bca-8dbf-e7940b1e4d96",
-    albumPosition: 3,
-    slug: "magma",
-    title: "Magma",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  lava: {
-    id: "3176c13f-78fc-40bd-8c1a-31746de228e8",
-    albumPosition: 4,
-    slug: "lava",
-    title: "Lava",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  hellsItch: {
-    id: "281cf9a2-be33-4eee-8f1d-c1ce5fb1c007",
-    albumPosition: 5,
-    slug: "hells-itch",
-    title: "Hell's Itch",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  ironLung: {
-    id: "15a32d56-c560-4017-9e49-e000b644a55b",
-    albumPosition: 6,
-    slug: "iron-lung",
-    title: "Iron Lung",
-    albumId: seedAlbums.iceDeath.id,
-  },
-  gliese710: {
-    id: "58dec2c7-6aee-4238-b62a-e4d90fbcd6bf",
-    albumPosition: 7,
-    slug: "gliese-710",
-    title: "Gliese 710",
-    albumId: seedAlbums.iceDeath.id,
-  },
-
-  // ====================================================
-  // Laminated Denim
-  theLandBeforeTimeland: {
-    id: "26528ce3-3226-464f-861b-9f51e21e6398",
-    albumPosition: 1,
-    slug: "the-land-before-timeland",
-    title: "The Land Before Timeland",
-    albumId: seedAlbums.laminatedDenim.id,
-  },
-  hypertension: {
-    id: "5765ec19-0e59-479c-a62c-a019ac73ce28",
-    albumPosition: 2,
-    slug: "hypertension",
-    title: "Hypertension",
-    albumId: seedAlbums.laminatedDenim.id,
-  },
-
-  // ====================================================
-  // Changes
-  change: {
-    id: "8e9e04f6-fc88-4846-a5df-977bba3f36c9",
-    albumPosition: 1,
-    slug: "change",
-    title: "Change",
-    albumId: seedAlbums.changes.id,
-  },
-  hateDancin: {
-    id: "39da3ba1-62e6-4e8b-b785-c9efc2e2a1ca",
-    albumPosition: 2,
-    slug: "hate-dancin",
-    title: "Hate Dancin'",
-    albumId: seedAlbums.changes.id,
-  },
-  astroturf: {
-    id: "b9cc7829-00b0-4421-a3fd-a3a004890ca2",
-    albumPosition: 3,
-    slug: "astroturf",
-    title: "Astroturf",
-    albumId: seedAlbums.changes.id,
-  },
-  noBody: {
-    id: "9b2d99e9-31f4-42a7-84a8-f89d22530e05",
-    albumPosition: 4,
-    slug: "no-body",
-    title: "No Body",
-    albumId: seedAlbums.changes.id,
-  },
-  gondii: {
-    id: "a2737d99-cecd-49c5-bdce-163057a28263",
-    albumPosition: 5,
-    slug: "gondii",
-    title: "Gondii",
-    albumId: seedAlbums.changes.id,
-  },
-  explodingSuns: {
-    id: "f0ec5bd9-0ce5-4449-a3a1-df41a5f2aecd",
-    albumPosition: 6,
-    slug: "exploding-suns",
-    title: "Exploding Suns",
-    albumId: seedAlbums.changes.id,
-  },
-  shortChange: {
-    id: "f4f841b3-2eb8-428a-82ea-ad73165ab35d",
-    albumPosition: 7,
-    slug: "short-change",
-    title: "Short Change",
-    albumId: seedAlbums.changes.id,
-  },
-
-  // ====================================================
-  // PetroDragonic Apocalypse; or, Dawn of Eternal Night:
-  // An Annihilation of Planet Earth and the Beginning of Merciless Damnation
-  motorSpirit: {
-    id: "7a076db2-b2b7-4cea-a0cb-2abdaf36c7a1",
-    albumPosition: 1,
-    slug: "motor-spirit",
-    title: "Motor Spirit",
-    albumId: seedAlbums.petro.id,
-  },
-  supercell: {
-    id: "eb1c3c7a-c735-4013-8408-64be214d6a39",
-    albumPosition: 2,
-    slug: "supercell",
-    title: "Supercell",
-    albumId: seedAlbums.petro.id,
-  },
-  converge: {
-    id: "8879d2b7-57e3-4e38-bc0c-250f3b5fb288",
-    albumPosition: 3,
-    slug: "converge",
-    title: "Converge",
-    albumId: seedAlbums.petro.id,
-  },
-  witchcraft: {
-    id: "e5e33a88-2ae2-4c49-914b-837d01aa088c",
-    albumPosition: 4,
-    slug: "witchcraft",
-    title: "Witchcraft",
-    albumId: seedAlbums.petro.id,
-  },
-  gilaMonster: {
-    id: "259d79c1-2e33-478f-b670-4e64d416c6f2",
-    albumPosition: 5,
-    slug: "gila-monster",
-    title: "Gila Monster",
-    albumId: seedAlbums.petro.id,
-  },
-  dragon: {
-    id: "615908d3-72ae-4034-8d28-59cdf0341967",
-    albumPosition: 6,
-    slug: "dragon",
-    title: "Dragon",
-    albumId: seedAlbums.petro.id,
-  },
-  flamethrower: {
-    id: "20a76786-5937-4b04-bcaa-37916a9edced",
-    albumPosition: 7,
-    slug: "flamethrower",
-    title: "Flamethrower",
-    albumId: seedAlbums.petro.id,
-  },
-
-  // ====================================================
-  // The Silver Cord
-  theia: {
-    id: "cadd1e1b-db88-439b-be3a-7dc402075636",
-    albumPosition: 1,
-    slug: "theia",
-    title: "Theia",
-    albumId: seedAlbums.tsc.id,
-  },
-  theSilverCord: {
-    id: "c7633144-0d11-45a6-ac88-ee089b7e9605",
-    albumPosition: 2,
-    slug: "the-silver-cord",
-    title: "The Silver Cord",
-    albumId: seedAlbums.tsc.id,
-  },
-  set: {
-    id: "d28b90db-4e89-48da-ace5-d311e82c550e",
-    albumPosition: 3,
-    slug: "set",
-    title: "Set",
-    albumId: seedAlbums.tsc.id,
-  },
-  "chang-e": {
-    id: "92052f74-b55f-404f-be92-4f68440db27f",
-    albumPosition: 4,
-    slug: "chang-e",
-    title: "Chang'e",
-    albumId: seedAlbums.tsc.id,
-  },
-  gilgamesh: {
-    id: "42ba1e70-c9cd-49b9-aa02-ab2e3e403e4c",
-    albumPosition: 5,
-    slug: "gilgamesh",
-    title: "Gilgamesh",
-    albumId: seedAlbums.tsc.id,
-  },
-  swanSong: {
-    id: "159f483a-85d1-4480-a4c6-3526f579e34f",
-    albumPosition: 6,
-    slug: "swan-song",
-    title: "Swan Song",
-    albumId: seedAlbums.tsc.id,
-  },
-  extinction: {
-    id: "5bd9512b-9dca-43c3-b6d1-2cda9221a7bb",
-    albumPosition: 7,
-    slug: "extinction",
-    title: "Extinction",
-    albumId: seedAlbums.tsc.id,
-  },
-
-  // ====================================================
-  // Flight b741
-  mirageCity: {
-    id: "c842a0d7-2f49-4a2d-8f05-4f2f19aae3e9",
-    albumPosition: 1,
-    slug: "mirage-city",
-    title: "Mirage City",
-    albumId: seedAlbums.b741.id,
-  },
-  antarctica: {
-    id: "9fef82b7-59e3-4471-8f17-5ab2cb5a3104",
-    albumPosition: 2,
-    slug: "antarctica",
-    title: "Antarctica",
-    albumId: seedAlbums.b741.id,
-  },
-  rawFeel: {
-    id: "c9550f02-d46d-4ba7-a595-1bf1bcf00262",
-    albumPosition: 3,
-    slug: "raw-feel",
-    title: "Raw Feel",
-    albumId: seedAlbums.b741.id,
-  },
-  fieldOfVision: {
-    id: "89b0330f-92a1-4df9-b3a9-539fabb5ab8a",
-    albumPosition: 4,
-    slug: "field-of-vision",
-    title: "Field of Vision",
-    albumId: seedAlbums.b741.id,
-  },
-  hogCallingContest: {
-    id: "a07d9124-08c3-49d5-83e4-f2d1806bbb34",
-    albumPosition: 5,
-    slug: "hog-calling-contest",
-    title: "Hog Calling Contest",
-    albumId: seedAlbums.b741.id,
-  },
-  leRisque: {
-    id: "98521dd4-8ca6-48e2-b7ed-05e3bc765711",
-    albumPosition: 6,
-    slug: "le-risque",
-    title: "Le Risque",
-    albumId: seedAlbums.b741.id,
-  },
-  flightb741: {
-    id: "0a71eafa-69c2-444a-b35a-259149e85937",
-    albumPosition: 7,
-    slug: "flight-b741",
-    title: "Flight b741",
-    albumId: seedAlbums.b741.id,
-  },
-  sadPilot: {
-    id: "dff597cf-44c3-4c0a-9219-ad7ff249cd26",
-    albumPosition: 8,
-    slug: "sad-pilot",
-    title: "Sad Pilot",
-    albumId: seedAlbums.b741.id,
-  },
-  ratsInTheSky: {
-    id: "37922a94-85c4-4a13-8155-de1ee68872ac",
-    albumPosition: 9,
-    slug: "rats-in-the-sky",
-    title: "Rats In The Sky",
-    albumId: seedAlbums.b741.id,
-  },
-  dailyBlues: {
-    id: "4da6d803-c03e-4d29-9f43-f9042b1cabcd",
-    albumPosition: 10,
-    slug: "daily-blues",
-    title: "Daily Blues",
-    albumId: seedAlbums.b741.id,
-  },
-} as const satisfies Record<string, Omit<Song, "createdAt">>;
-export const allSongs = Object.values(seedSongs) as Song[];
-
-// =============================================================================
-// ================================= SHOWS =====================================
-// =============================================================================
-
-// Ordered by show date
-export const seedShows = {
-  // ====================================================
-  // 2013 Tour
-  northcote13: {
-    id: "86907172-176f-4930-8017-20e1b044d142",
-    slug: "northcote-2013",
-    location: "Melbourne",
-    date: "2013-08-29",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-
-  // ====================================================
-  // 2016 Tour
-  sf16: {
-    id: "959fc957-a47f-4618-a0ea-91a57e17c55e",
-    slug: "san-francisco-2016",
-    location: "San Francisco",
-    date: "2016-05-25",
-    imageUrl: "https://f4.bcbits.com/img/a3583868136_16.jpg",
-    bandcampAlbumId: "1532371660",
-  },
-
-  // ====================================================
-  // 2018 Tour
-  kexp18: {
-    id: "fd001d84-9b97-4525-ac2b-6f9cfece6136",
-    slug: "kexp-2018",
-    location: "KEXP",
-    date: "2018-06-06",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-
-  // ====================================================
-  // World Tour 2019
-  adelaide19: {
-    id: "3c665da1-5458-4b1e-bcae-ab063e72385f",
-    slug: "adelaide-2019",
-    location: "Adelaide",
-    date: "2019-07-12",
-    imageUrl: "https://f4.bcbits.com/img/a0297850511_16.jpg",
-    bandcampAlbumId: "1286108847",
-  },
-  milwaukee19: {
-    id: "ccb86577-5c06-4180-8ff1-781c7c41e8d0",
-    slug: "milwaukee-2019",
-    location: "Milwaukee",
-    date: "2019-08-23",
-    imageUrl: "https://f4.bcbits.com/img/a2134759453_16.jpg",
-    bandcampAlbumId: "2354695487",
-  },
-  manchester19: {
-    id: "01575133-ff92-4668-99d4-b1b62a9f29eb",
-    slug: "manchester-2019",
-    location: "Manchester",
-    date: "2019-10-03",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-  brussels19: {
-    id: "493db3ef-ee80-4e12-acc2-6fb046b79408",
-    slug: "brussels-2019",
-    location: "Brussels",
-    date: "2019-10-08",
-    imageUrl: "https://f4.bcbits.com/img/a2684291974_16.jpg",
-    bandcampAlbumId: "1602825616",
-  },
-  madrid19: {
-    id: "11597d9b-21e7-43b7-b681-cd22a31c5a90",
-    slug: "madrid-2019",
-    location: "Madrid",
-    date: "2019-10-18",
-    imageUrl: "https://f4.bcbits.com/img/a3650674791_16.jpg",
-    bandcampAlbumId: "6509641",
-  },
-
-  // ====================================================
-  // Micro Tour 2021
-  sydney21: {
-    id: "95a0d2a2-4147-44b5-9f4c-ef871a13bf62",
-    slug: "sydney-2021",
-    location: "Sydney",
-    date: "2021-04-22",
-    imageUrl: "https://f4.bcbits.com/img/a0572452225_16.jpg",
-    bandcampAlbumId: "2609270300",
-  },
-
-  // ====================================================
-  // 2021 Princess Theatre Residency
-  brisbaneAcoustic21: {
-    id: "81c86246-d8e8-4023-9be9-c72042ffdefe",
-    slug: "brisbane-acoustic-2021",
-    location: "Brisbane",
-    date: "2021-12-19",
-    imageUrl: "https://f4.bcbits.com/img/a2213415853_16.jpg",
-    bandcampAlbumId: "1173053054",
-  },
-
-  // ====================================================
-  // World Tour 2022
-  melbourne22: {
-    id: "4fc59426-d7bf-4726-9270-d76630b7569f",
-    slug: "melbourne-2022",
-    location: "Melbourne", // Timeland
-    date: "2022-04-05",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-  pomona22: {
-    id: "24f4df1a-7921-4ab0-a618-0479d4a8b2d7",
-    slug: "pomona-2022",
-    location: "Pomona",
-    date: "2022-04-12",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-  slo22: {
-    id: "cfbee1cf-66d6-4cf7-b539-ab1a56d4656f",
-    slug: "slo-2022",
-    location: "San Luis Obispo",
-    date: "2022-04-24",
-    imageUrl: "/showCovers/slo22.webp",
-    bandcampAlbumId: null,
-  },
-  mexicoCity22: {
-    id: "8c4b2700-ed44-4a14-8d08-93f6c7a0b5a3",
-    slug: "mexico-city-2022",
-    location: "Mexico City",
-    date: "2022-05-06",
-    imageUrl: "/showCovers/mexicoCity22.webp",
-    bandcampAlbumId: null,
-  },
-  bonnaroo22: {
-    id: "1744f3c5-8da4-4e28-845d-0b2cf02db7a1",
-    slug: "bonnaroo-2022",
-    location: "Bonnaroo",
-    date: "2022-06-17",
-    imageUrl: "https://f4.bcbits.com/img/a3553224286_16.jpg",
-    bandcampAlbumId: "1363173846",
-  },
-  croatia22: {
-    id: "c0c2fe6d-92b8-4152-923c-e22fa18570e2",
-    slug: "croatia-2022",
-    location: "Croatia",
-    date: "2022-08-02",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-  berkeley22: {
-    id: "cb025e04-05bb-4569-8797-9d58d35e21ee",
-    slug: "berkeley-2022",
-    location: "Berkeley",
-    date: "2022-10-02",
-    imageUrl: "/showCovers/berkeley22.webp",
-    bandcampAlbumId: null,
-  },
-  rr22: {
-    id: "ca4d9095-cb3a-46d9-8dd8-3fdc7b15a7b6",
-    slug: "red-rocks-2022",
-    location: "Red Rocks",
-    date: "2022-10-10",
-    imageUrl: "https://f4.bcbits.com/img/a3180996672_16.jpg",
-    bandcampAlbumId: "2418689883",
-  },
-  philadelphia22: {
-    id: "f5ab804d-be48-4de5-b096-ceb33f1551fd",
-    slug: "philadelphia-2022",
-    location: "Philadelphia",
-    date: "2022-10-22",
-    imageUrl: "/showCovers/philadelphia22.webp",
-    bandcampAlbumId: null,
-  },
-
-  // ====================================================
-  // U.S. Residency Tour 2023
-  caverns23: {
-    id: "edb045a8-3998-42bd-8a8a-540b8124e3e3",
-    slug: "the-caverns-2023",
-    location: "The Caverns",
-    date: "2023-06-01",
-    imageUrl: "/showCovers/caverns23.webp",
-    bandcampAlbumId: null,
-  },
-  rr23: {
-    id: "af055b78-9c40-4f15-a676-ed05f3485019",
-    slug: "red-rocks-2023",
-    location: "Red Rocks",
-    date: "2023-06-07",
-    imageUrl: "/showCovers/rr23.webp",
-    bandcampAlbumId: null,
-  },
-  chicago23: {
-    id: "1f8192d8-4650-4837-b119-b82507afe23c",
-    slug: "chicago-2023",
-    location: "Chicago",
-    date: "2023-06-11",
-    imageUrl: "https://f4.bcbits.com/img/a1324413369_16.jpg",
-    bandcampAlbumId: "3977429038",
-  },
-  remlinger23: {
-    id: "ae71a1cd-0823-4feb-85f8-065529415dff",
-    slug: "remlinger-farms-2023",
-    location: "Remlinger Farms",
-    date: "2023-06-17",
-    imageUrl: "/showCovers/remlinger23.webp",
-    bandcampAlbumId: null,
-  },
-  hollywoodBowl23: {
-    id: "cb9a091f-d5e3-42b1-84ce-d9a5e72a7ff6",
-    slug: "hollywood-bowl-2023",
-    location: "Hollywood Bowl",
-    date: "2023-06-21",
-    imageUrl: "/showCovers/hollywoodBowl23.webp",
-    bandcampAlbumId: null,
-  },
-
-  // ====================================================
-  // Europe & UK Tour 2023
-  warsaw23: {
-    id: "4c9a09b8-d363-4220-976e-1071ab0b302c",
-    slug: "warsaw-2023",
-    location: "Warsaw",
-    date: "2023-03-11",
-    imageUrl: "/showCovers/warsaw23.webp",
-    bandcampAlbumId: null,
-  },
-  palp23: {
-    id: "12a96ecf-2077-4264-8278-e76865b7507e",
-    slug: "palp-2023",
-    location: "Palp",
-    date: "2023-08-14",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-  lowlands23: {
-    id: "58a8f68b-0f95-4a3a-963b-9740a68b8729",
-    slug: "lowlands-2023",
-    location: "Lowlands",
-    date: "2023-08-19",
-    imageUrl: null,
-    bandcampAlbumId: null,
-  },
-
-  // ====================================================
-  // U.S. Tour 2024 pt. 1
-  dc24: {
-    id: "1aba10fe-3d4e-49c4-ae47-f30a2b7e0e1b",
-    slug: "dc-2024",
-    location: "D.C.",
-    date: "2024-08-15",
-    imageUrl: "https://f4.bcbits.com/img/a1586375238_16.jpg",
-    bandcampAlbumId: "2799349219",
-  },
-  nyc24: {
-    id: "4023c755-247d-4b29-86ec-7b1203e93bdf",
-    slug: "nyc-2024",
-    location: "New York City",
-    date: "2024-08-16",
-    imageUrl: "https://f4.bcbits.com/img/a4160066698_16.jpg",
-    bandcampAlbumId: "3530855413",
-  },
-  boston24: {
-    id: "dd72f78a-23c4-46fc-a82a-6befe9eac3e7",
-    slug: "boston-2024",
-    location: "Boston",
-    date: "2024-08-19",
-    imageUrl: "https://f4.bcbits.com/img/a4053913678_16.jpg",
-    bandcampAlbumId: "2151609245",
-  },
-  maine24: {
-    id: "212bb7b3-2ed9-4270-bf35-7026716a36b6",
-    slug: "maine-2024",
-    location: "Maine",
-    date: "2024-08-20",
-    imageUrl: "https://f4.bcbits.com/img/a1001399634_16.jpg",
-    bandcampAlbumId: "2637870866",
-  },
-  toronto24: {
-    id: "4b31bce2-5b79-4ea6-91ea-97b7d0a47aae",
-    slug: "toronto-2024",
-    location: "Toronto",
-    date: "2024-08-21",
-    imageUrl: "https://f4.bcbits.com/img/a1394424151_16.jpg",
-    bandcampAlbumId: "3363110237",
-  },
-  detroit24: {
-    id: "3bccdb81-d010-47e8-a9f8-4eb87f3f4066",
-    slug: "detroit-2024",
-    location: "Detroit",
-    date: "2024-08-23",
-    imageUrl: "https://f4.bcbits.com/img/a0824108206_16.jpg",
-    bandcampAlbumId: "1675633392",
-  },
-  cleveland24: {
-    id: "491b11cb-b085-4081-a6d0-a9070532b82d",
-    slug: "cleveland-2024",
-    location: "Cleveland",
-    date: "2024-08-24",
-    imageUrl: "https://f4.bcbits.com/img/a4062916790_16.jpg",
-    bandcampAlbumId: "4147847254",
-  },
-  kentucky24: {
-    id: "fdc39520-041b-4faf-bde8-8288dbbf4f55",
-    slug: "kentucky-2024",
-    location: "Kentucky",
-    date: "2024-08-25",
-    imageUrl: "https://f4.bcbits.com/img/a0274986703_16.jpg",
-    bandcampAlbumId: "1304619880",
-  },
-  philadelphia24: {
-    id: "3924bba7-a859-4ce0-808c-2eacdf93b4bf",
-    slug: "philadelphia-2024",
-    location: "Philadelphia",
-    date: "2024-08-27",
-    imageUrl: "https://f4.bcbits.com/img/a2253820167_16.jpg",
-    bandcampAlbumId: "1839675933",
-  },
-  richmond24: {
-    id: "b9acb3f0-a406-44be-8fb5-26aedf47ee61",
-    slug: "richmond-2024",
-    location: "Richmond",
-    date: "2024-08-28",
-    imageUrl: "https://f4.bcbits.com/img/a4074051612_16.jpg",
-    bandcampAlbumId: "1275144158",
-  },
-  asheville24: {
-    id: "704abe4c-3cb2-4391-9a81-a5dcb1fd3a44",
-    slug: "asheville-2024",
-    location: "Asheville",
-    date: "2024-08-30",
-    imageUrl: "https://f4.bcbits.com/img/a2066920903_16.jpg",
-    bandcampAlbumId: "817110471",
-  },
-  nashville24: {
-    id: "f4b93c00-1998-4c3d-b625-f55d2e6b426c",
-    slug: "nashville-2024",
-    location: "Nashville",
-    date: "2024-08-31",
-    imageUrl: "https://f4.bcbits.com/img/a0545000465_16.jpg",
-    bandcampAlbumId: "3239781259",
-  },
-  chicago24: {
-    id: "6556aa55-88d3-43f8-bfdd-f348352ae1d0",
-    slug: "chicago-2024",
-    location: "Chicago",
-    date: "2024-09-01",
-    imageUrl: "https://f4.bcbits.com/img/a1788111107_16.jpg",
-    bandcampAlbumId: "4294151713",
-  },
-  minneapolis24: {
-    id: "04489d98-d04c-4980-bd62-b9f04807eb9f",
-    slug: "minneapolis-2024",
-    location: "Minneapolis",
-    date: "2024-09-03",
-    imageUrl: "https://f4.bcbits.com/img/a2807501952_16.jpg",
-    bandcampAlbumId: "1631866248",
-  },
-  milwaukee24: {
-    id: "6aadd547-499e-4b52-9b29-76682af9614e",
-    slug: "milwaukee-2024",
-    location: "Milwaukee",
-    date: "2024-09-04",
-    imageUrl: "https://f4.bcbits.com/img/a0203986788_16.jpg",
-    bandcampAlbumId: "1407410068",
-  },
-  stl24: {
-    id: "29361dc0-6671-4297-8157-5eb03bdbf7b2",
-    slug: "st-louis-2024",
-    location: "St. Louis",
-    date: "2024-09-05",
-    imageUrl: "https://f4.bcbits.com/img/a1541766080_16.jpg",
-    bandcampAlbumId: "2447524963",
-  },
-  rr24: {
-    id: "8fc370f6-46e7-488f-8e69-e4e3f3d10d28",
-    slug: "red-rocks-2024",
-    location: "Red Rocks",
-    date: "2024-09-08",
-    imageUrl: "https://f4.bcbits.com/img/a3263389792_16.jpg",
-    bandcampAlbumId: "79615285",
-  },
-  oregon24: {
-    id: "1e8be424-f5cf-4d4a-b6c2-b124fdeb405f",
-    slug: "oregon-2024",
-    location: "Oregon",
-    date: "2024-09-11",
-    imageUrl: "https://f4.bcbits.com/img/a1151612533_10.jpg",
-    bandcampAlbumId: "2716193401",
-  },
-  vancouver24: {
-    id: "d5147b82-c83b-448c-9d7f-9b2113da8cf7",
-    slug: "vancouver-2024",
-    location: "Vancouver",
-    date: "2024-09-12",
-    imageUrl: "https://f4.bcbits.com/img/a2167864021_16.jpg",
-    bandcampAlbumId: "4272985994",
-  },
-  gorge24: {
-    id: "69af662b-164e-499e-a459-e9cb903aae97",
-    slug: "the-gorge-2024",
-    location: "The Gorge",
-    date: "2024-09-14",
-    imageUrl: "https://f4.bcbits.com/img/a2893046862_16.jpg",
-    bandcampAlbumId: "3746606297",
-  },
-
-  // ====================================================
-  // U.S. Tour 2024 pt. 2
-  la24: {
-    id: "a81ef768-32b5-4fbd-8d9e-fefd3cf44411",
-    slug: "los-angeles-2024",
-    location: "Los Angeles",
-    date: "2024-11-01",
-    imageUrl: "https://f4.bcbits.com/img/a3083529929_16.jpg",
-    bandcampAlbumId: "2053666447",
-  },
-  sd24: {
-    // Acoustic!
-    id: "481aec4e-39bb-48d9-8082-4469c74e00da",
-    slug: "san-diego-2024",
-    location: "San Diego",
-    date: "2024-11-02",
-    imageUrl: "https://f4.bcbits.com/img/a3464482652_16.jpg",
-    bandcampAlbumId: "3483464958",
-  },
-  pasoRobles24: {
-    id: "372c1f0a-f676-46a3-9f21-34fdf02c30c5",
-    slug: "paso-robles-2024",
-    location: "Paso Robles",
-    date: "2024-11-03",
-    imageUrl: "https://f4.bcbits.com/img/a3550028268_16.jpg",
-    bandcampAlbumId: "3875461640",
-  },
-  stanford24: {
-    id: "935dd6ed-2153-4e4b-8bea-1d75825b8fe4",
-    slug: "stanford-2024",
-    location: "Stanford",
-    date: "2024-11-04",
-    imageUrl: "https://f4.bcbits.com/img/a2511383545_16.jpg",
-    bandcampAlbumId: "2620217443",
-  },
-  sf24: {
-    id: "0755f77f-3f36-498e-99b0-5f71228076a8",
-    slug: "san-francisco-2024",
-    location: "San Francisco",
-    date: "2024-11-06",
-    imageUrl: "https://f4.bcbits.com/img/a2123423689_16.jpg",
-    bandcampAlbumId: "783019413",
-  },
-  vegas24: {
-    id: "06c550a5-028e-438c-b8a7-6ede67b3737f",
-    slug: "las-vegas-2024",
-    location: "Las Vegas",
-    date: "2024-11-08",
-    imageUrl: "https://f4.bcbits.com/img/a3468174213_16.jpg",
-    bandcampAlbumId: "1104627077",
-  },
-  phoenix24: {
-    id: "8bd9f809-7ab8-4046-9347-e349eb663b86",
-    slug: "phoenix-2024",
-    location: "Phoenix",
-    date: "2024-11-09",
-    imageUrl: "https://f4.bcbits.com/img/a3527600251_16.jpg",
-    bandcampAlbumId: "4200479984",
-  },
-  abq24: {
-    id: "f0139f08-7075-4757-b811-9c458d04593c",
-    slug: "albuquerque-2024",
-    location: "Albuquerque",
-    date: "2024-11-10",
-    imageUrl: "https://f4.bcbits.com/img/a0496185183_16.jpg",
-    bandcampAlbumId: "660285328",
-  },
-  okc24: {
-    id: "6070a189-e1a7-42ce-8d19-f5e43921967d",
-    slug: "oklahoma-city-2024",
-    location: "Oklahoma City",
-    date: "2024-11-12",
-    imageUrl: "https://f4.bcbits.com/img/a1910276487_16.jpg",
-    bandcampAlbumId: "2847156728",
-  },
-  arkansas24: {
-    id: "a647481f-4159-4351-9d9d-53630646a3dc",
-    slug: "arkansas-2024",
-    location: "Arkansas",
-    date: "2024-11-13",
-    imageUrl: "https://f4.bcbits.com/img/a3844836167_16.jpg",
-    bandcampAlbumId: "2294934601",
-  },
-  austin24: {
-    id: "2b1b2dec-7264-4891-ba73-f7116e0818ff",
-    slug: "austin-2024",
-    location: "Austin",
-    date: "2024-11-15",
-    imageUrl: "https://f4.bcbits.com/img/a2463924603_16.jpg",
-    bandcampAlbumId: "719227306",
-  },
-  houston24: {
-    id: "8b1615e0-a6dc-42f6-b61c-465971545450",
-    slug: "houston-2024",
-    location: "Houston",
-    date: "2024-11-16",
-    imageUrl: "https://f4.bcbits.com/img/a3220456230_16.jpg",
-    bandcampAlbumId: "3854890243",
-  },
-  nola24: {
-    id: "48d267dc-90a2-401d-aec5-b127a7abe06d",
-    slug: "new-orleans-2024",
-    location: "New Orleans",
-    date: "2024-11-17",
-    imageUrl: "https://f4.bcbits.com/img/a2503962224_16.jpg",
-    bandcampAlbumId: "3455343812",
-  },
-  atlanta24: {
-    id: "8af52ec8-f9a6-43e3-ad24-a5ac42f3308d",
-    slug: "atlanta-2024",
-    location: "Atlanta",
-    date: "2024-11-19",
-    imageUrl: "https://f4.bcbits.com/img/a2812384254_16.jpg",
-    bandcampAlbumId: "3316613276",
-  },
-  stAugustine24: {
-    id: "2abeb068-9075-483d-8dc3-66a24a5e528c",
-    slug: "st-augustine-2024",
-    location: "St. Augustine",
-    date: "2024-11-20",
-    imageUrl: "https://f4.bcbits.com/img/a1954948404_16.jpg",
-    bandcampAlbumId: "1834138981",
-  },
-  miami24: {
-    id: "2c2ae0d1-f46a-409e-9fa4-1c046bdacd9d",
-    slug: "miami-2024",
-    location: "Miami",
-    date: "2024-11-21",
-    imageUrl: "https://f4.bcbits.com/img/a1675499924_16.jpg",
-    bandcampAlbumId: "4243074642",
-  },
-} as const satisfies Record<string, Omit<Show, "createdAt">>;
-export const allShows = Object.values(seedShows) as Show[];
-
-// =============================================================================
-// ============================= PERFROMANCES ==================================
+// ============================= Performances ==================================
 // =============================================================================
 
 // Ordered alphabetically by song title, not ordered with songs
@@ -5793,4 +3316,5 @@ export const seedPerformances = [
   Performance,
   "createdAt" | "ratingLastUpdatedAt" | "eloRating"
 >[];
+
 export const allPerformances = seedPerformances as unknown as Performance[];
